Add supervisor and user routes for managing accounts and certificates
- Implemented routes for supervisor account creation, login, and dashboard management.
- Added functionality for handling organization link requests, including approval and rejection.
- Created user routes for managing uploaded certificates, including upload, update, delete, and sharing.
- Enhanced session management and CSRF protection across all routes.
- Introduced structured error handling and user feedback for various actions.
