+++ app.py
@@ -3,72 +3,59 @@
 
 from __future__ import annotations
 
-from calendar import monthrange
-from datetime import date, timedelta
-import atexit
-from functools import partial
-import importlib
-import importlib.util
 import logging
 import os
-import re
-import secrets
-import sys
-from pathlib import Path
-import time
-from typing import Any, Callable, Sequence
-import json
 
-from flask import (
-    Flask,
-    Response,
-    abort,
-    current_app,
-    flash,
-    g,
-    jsonify,
-    send_from_directory,
-    make_response,
-    redirect,
-    render_template,
-    request,
-    session,
-    url_for,
-)
-from flask.typing import ResponseReturnValue
-from markupsafe import Markup
-from werkzeug.exceptions import HTTPException, RequestEntityTooLarge
-from werkzeug.middleware.proxy_fix import ProxyFix
+from flask import Flask
 
-from config_loader import load_environment
-from functions.logging import (
-    configure_module_logger,
-    configure_root_logging,
-    mask_email,
-    mask_email_reference,
-    mask_hash,
-    mask_headers,
-    mask_sensitive_data,
-)
+from web import bootstrap
 
-from course_categories import (
-    COURSE_CATEGORIES,
-    COURSE_CATEGORY_GROUPS,
-    labels_for_slugs,
-    normalize_category_slugs,
-)
-
-from functions.emails import service as email_service
-from functions.pdf import service as pdf
-from functions.notifications import critical_events
-from functions import security as sec
-from functions.requests import as_bool, get_request_ip, register_public_submission
-
-
-load_environment()
-configure_root_logging()
+bootstrap.initialize_runtime()
 
 import functions
+from functions import security as sec
+from functions.emails import service as email_service
+from functions.logging import configure_module_logger
+from functions.notifications import critical_events
+from functions.pdf import service as pdf
+from functions.requests import as_bool, get_request_ip, register_public_submission
+
+from web.errors import register_error_handlers, request_entity_too_large
+from web.helpers import (
+    ALLOWED_ADMIN_APPROVAL_ERRORS,
+    ALLOWED_ADMIN_REJECTION_ERRORS,
+    ALLOWED_PDF_METADATA_UPDATE_ERRORS,
+    ALLOWED_PDF_UPLOAD_ERRORS,
+    ALLOWED_SUPERVISOR_ACTIVATION_ERRORS,
+    CLIENT_LOG_TRUNCATION_LIMITS,
+    CSRF_EXPIRED_MESSAGE,
+    TOO_MANY_ATTEMPTS_MESSAGE,
+    UPLOAD_MAX_BYTES,
+    UPLOAD_MAX_MB,
+    UPLOAD_TOO_LARGE_MESSAGE,
+    SafeUserPayloadError,
+    _add_months_to_date,
+    _api_error_response,
+    _coerce_text_payload_value,
+    _editable_pdf_name,
+    _format_display_name,
+    _log_api_error,
+    _mask_sensitive_fields,
+    _mask_username_for_log,
+    _parse_duration_expiry_value,
+    _render_create_supervisor_page,
+    _request_error_context,
+    _resolve_certificate_expiry,
+    _safe_user_error,
+    _sanitize_search_term,
+    _truncate_log_value,
+    render_markdown_content,
+)
+from web.routes_admin import register_admin_routes
+from web.routes_public import register_public_routes
+from web.routes_supervisor import register_supervisor_routes
+from web.routes_user import register_user_routes
+
 
 ensure_csrf_token = sec.ensure_csrf_token
 validate_csrf_token = sec.validate_csrf_token
@@ -78,478 +65,47 @@ save_pdf_for_user = pdf.save_pdf_for_user
 logger = configure_module_logger(__name__)
 
 
-def _render_create_supervisor_page(error: str | None = None, invalid: bool = False, **extra) -> str:
-    """Returnerar render_template för create_supervisor/create_user-sidor med standard-kwargs.
-
-    Detta förenklar upprepade anrop som tidigare skickade samma kwargs flera gånger.
-    """
-    common = {
-        "invalid": invalid,
-        "page_title": "Skapa konto",
-        "heading": "Skapa konto",
-        "description": (
-            "Välj ett starkt lösenord för ditt konto. "
-            "Lösenordet måste vara minst åtta tecken långt."
-        ),
-        "submit_text": "Skapa konto",
-    }
-    common.update(extra or {})
-    if error:
-        common["error"] = error
-    return render_template("create_supervisor.html", **common)
-
-
-def _render_basic_markdown(text: str) -> str:
-    rendered_lines = []
-    for line in text.splitlines():
-        stripped = line.strip()
-        if stripped.startswith("## "):
-            rendered_lines.append(f"<h2>{stripped[3:].strip()}</h2>")
-            continue
-        if stripped.startswith("# "):
-            rendered_lines.append(f"<h1>{stripped[2:].strip()}</h1>")
-            continue
-        if stripped:
-            rendered_lines.append(f"<p>{stripped}</p>")
-    return "\n".join(rendered_lines)
-
-
-def render_markdown_content(text: str) -> Markup:
-    markdown_spec = importlib.util.find_spec("markdown")
-    if markdown_spec is None:
-        logger.warning("Markdown-biblioteket saknas, använder enkel renderare.")
-        return Markup(_render_basic_markdown(text))
-    markdown_module = importlib.import_module("markdown")
-    rendered = markdown_module.markdown(text, extensions=["extra"])
-    return Markup(rendered)
-
-
-ALLOWED_PDF_UPLOAD_ERRORS = {
-    "Ingen fil vald.",
-    "Endast PDF, PNG eller JPG tillåts.",
-    "Filen verkar inte vara en giltig PDF.",
-    "Exakt en kurskategori måste väljas.",
-    "Bilden kunde inte konverteras till PDF.",
-    "PDF:en blockerades av säkerhetsskannern.",
-}
-ALLOWED_PDF_METADATA_UPDATE_ERRORS = {
-    "Intygsnamnet kan inte vara tomt.",
-    "Intygsnamnet innehåller inga tillåtna tecken.",
-    "Intygsnamnet får vara högst 120 tecken.",
-    "Intygsnamnet måste anges som text.",
-    "Anteckningen måste anges som text.",
-    "Utgångsdatum måste anges som text.",
-    "Antal månader måste anges som text.",
-    "Antal år måste anges som text.",
-    "Välj ett utgångsdatum.",
-    "Välj ett giltigt utgångsdatum.",
-    "Utgångsdatum kan inte vara tidigare än idag.",
-    "Ange ett giltigt antal månader.",
-    "Antal månader kan inte vara mindre än 0.",
-    "Antal månader får vara högst 1200.",
-    "Ange ett giltigt antal år.",
-    "Antal år kan inte vara mindre än 0.",
-    "Antal år får vara högst 100.",
-    "Ange antal år, månader eller båda.",
-    "Välj ett giltigt alternativ för utgångsdatum.",
-}
-ALLOWED_SUPERVISOR_ACTIVATION_ERRORS = {
-    "Lösenordet måste vara minst 8 tecken.",
-}
-ALLOWED_ADMIN_APPROVAL_ERRORS = {
-    "Ansökan hittades inte.",
-    "Ansökan är redan hanterad.",
-    "Ansökan saknar personnummer och kan inte godkännas.",
-    "E-postadressen är redan registrerad.",
-    "Företagsnamn saknas för detta organisationsnummer.",
-}
-ALLOWED_ADMIN_REJECTION_ERRORS = {
-    "Ansökan hittades inte.",
-    "Ansökan är redan hanterad.",
-}
-
-CLIENT_LOG_TRUNCATION_LIMITS = {
-    "message": 500,
-    "context": 200,
-    "url": 500,
-    "details": 1000,
-}
-
-UPLOAD_MAX_MB = 50
-UPLOAD_MAX_BYTES = 52_428_800 
-
-# Gemensamma användarmeddelanden
-CSRF_EXPIRED_MESSAGE = "Formuläret är inte längre giltigt. Ladda om sidan och försök igen."
-TOO_MANY_ATTEMPTS_MESSAGE = "Du har gjort för många försök. Vänta en stund och prova igen."
-UPLOAD_TOO_LARGE_MESSAGE = (
-    f"Uppladdningen är för stor. Max {UPLOAD_MAX_MB} MB tillåts."
-)
-
-
-class SafeUserPayloadError(ValueError):
-    # Markerar valideringsfel som redan är säkra att visa direkt för användaren.
-    pass
-
-
-def _safe_user_error(message: str, allowed: set[str], fallback: str) -> str:
-    # Returnera bara godkända felmeddelanden till användaren.
-    cleaned = (message or "").strip()
-    if cleaned in allowed:
-        return cleaned
-    return fallback
-
-
-def _add_months_to_date(base_date: date, months: int) -> date:
-    # Bevara dag i månaden när det går, annars använd månadens sista dag.
-    absolute_month = (base_date.year * 12) + (base_date.month - 1) + months
-    target_year, target_month_index = divmod(absolute_month, 12)
-    target_month = target_month_index + 1
-    target_day = min(base_date.day, monthrange(target_year, target_month)[1])
-    return date(target_year, target_month, target_day)
-
-
-def _parse_duration_expiry_value(raw_value: str, field_label: str, maximum: int) -> int:
-    cleaned = (raw_value or "").strip()
-    if not cleaned:
-        return 0
-    try:
-        parsed = int(cleaned)
-    except ValueError as exc:
-        raise ValueError(f"Ange ett giltigt antal {field_label}.") from exc
-    if parsed < 0:
-        raise ValueError(f"Antal {field_label} kan inte vara mindre än 0.")
-    if parsed > maximum:
-        raise ValueError(f"Antal {field_label} får vara högst {maximum}.")
-    return parsed
-
-
-def _resolve_certificate_expiry(
-    expiry_mode: str,
-    expiry_date_raw: str,
-    expiry_months_raw: str,
-    expiry_years_raw: str,
-    *,
-    current_expires_on: date | None = None,
-    today: date | None = None,
-) -> date | None:
-    today = today or date.today()
-    normalized_mode = (expiry_mode or "none").strip().lower()
-    if normalized_mode in {"", "none"}:
-        return None
-
-    if normalized_mode == "date":
-        cleaned_date = (expiry_date_raw or "").strip()
-        if not cleaned_date:
-            raise ValueError("Välj ett utgångsdatum.")
-        try:
-            expires_on = date.fromisoformat(cleaned_date)
-        except ValueError as exc:
-            raise ValueError("Välj ett giltigt utgångsdatum.") from exc
-        if current_expires_on is not None and expires_on == current_expires_on:
-            return expires_on
-        if expires_on < today:
-            raise ValueError("Utgångsdatum kan inte vara tidigare än idag.")
-        return expires_on
-
-    if normalized_mode in {"duration", "months", "years"}:
-        month_count = _parse_duration_expiry_value(
-            expiry_months_raw,
-            "månader",
-            1200,
-        )
-        year_count = _parse_duration_expiry_value(
-            expiry_years_raw,
-            "år",
-            100,
-        )
-        if normalized_mode == "months" and month_count == 0:
-            raise ValueError("Ange hur många månader intyget gäller.")
-        if normalized_mode == "years" and year_count == 0:
-            raise ValueError("Ange hur många år intyget gäller.")
-        if normalized_mode == "duration" and month_count == 0 and year_count == 0:
-            raise ValueError("Ange antal år, månader eller båda.")
-        return _add_months_to_date(today, month_count + (year_count * 12))
-
-    raise ValueError("Välj ett giltigt alternativ för utgångsdatum.")
-
-
-def _coerce_text_payload_value(value: Any, field_label: str, *, default: str = "") -> str:
-    if value is None:
-        return default
-    if isinstance(value, str):
-        return value
-    if isinstance(value, (int, float, bool)):
-        return str(value)
-    raise SafeUserPayloadError(f"{field_label} måste anges som text.")
-
-
-def _format_display_name(name: str | None) -> str:
-    # Säkerställ stor begynnelsebokstav för varje namnsegment.
-    cleaned_name = (name or "").strip()
-    if not cleaned_name:
-        return ""
-    return " ".join(part[:1].upper() + part[1:] for part in cleaned_name.split())
-
-
-def _editable_pdf_name(filename: str) -> str:
-    # Visa gamla timestamp-prefixade filnamn utan teknisk prefix i redigeringsfältet.
-    extracted = pdf.extract_editable_pdf_name(filename)
-    cleaned = re.sub(r"\.pdf$", "", extracted, flags=re.IGNORECASE).strip()
-    return cleaned or "intyg"
-
-
-def _request_error_context(extra: dict[str, Any] | None = None) -> dict[str, Any]:
-    # Samla korrelationsdata för enhetlig felloggning.
-    session_personnummer = session.get("personnummer")
-    masked_user = (
-        mask_hash(session_personnummer) if isinstance(session_personnummer, str) and session_personnummer else None
-    )
-    context: dict[str, Any] = {
-        "endpoint": request.path,
-        "method": request.method,
-        "admin": session.get("admin_username") if session.get("admin_logged_in") else None,
-        "user": masked_user,
-    }
-    if extra:
-        context.update(extra)
-    return context
-
-
-def _log_api_error(
-    *,
-    user_message: str,
-    status_code: int,
-    severity: str,
-    error: Exception | None = None,
-    extra: dict[str, Any] | None = None,
-) -> None:
-    # Standardiserad loggning av API-fel med korrelationsdata.
-    context = _request_error_context(extra)
-    if severity == "warning":
-        logger.warning(
-            "API-fel (%s): %s | context=%s | error=%s",
-            status_code,
-            user_message,
-            context,
-            str(error) if error else "-",
-        )
-        return
-    if error is not None:
-        logger.exception(
-            "API-systemfel (%s): %s | context=%s",
-            status_code,
-            user_message,
-            context,
-        )
-        return
-    logger.error("API-systemfel (%s): %s | context=%s", status_code, user_message, context)
-
-
-def _api_error_response(
-    user_message: str,
-    status_code: int,
-    *,
-    severity: str = "warning",
-    error: Exception | None = None,
-    extra: dict[str, Any] | None = None,
-) -> tuple[Response, int]:
-    # Standardiserad felrespons för JSON-API.
-    _log_api_error(
-        user_message=user_message,
-        status_code=status_code,
-        severity=severity,
-        error=error,
-        extra=extra,
-    )
-    return jsonify({"status": "error", "message": user_message}), status_code
-
-
-def _truncate_log_value(value: Any, limit: int) -> str:
-    # Begränsa loggsträngar för att undvika enorma loggar.
-    if value is None:
-        return ""
-    text = str(value).strip()
-    if len(text) <= limit:
-        return text
-    return f"{text[:limit]}…"
-
-
-_SENSITIVE_KEYS = frozenset(
-    {
-        "email",
-        "personnummer",
-        "ssn",
-        "name",
-        "namn",
-        "forename",
-        "surname",
-        "username",
-    }
-)
-
-
-def _mask_sensitive_fields(value: Any) -> Any:
-    # Maskera potentiellt känsliga fält i API-svar/loggar för adminverktyg.
-    if isinstance(value, dict):
-        masked: dict[str, Any] = {}
-        for key, item in value.items():
-            lowered = str(key).lower()
-            if lowered in _SENSITIVE_KEYS:
-                masked[key] = "***"
-            else:
-                masked[key] = _mask_sensitive_fields(item)
-        return masked
-    if isinstance(value, list):
-        return [_mask_sensitive_fields(item) for item in value]
-    return value
-
-
-def _sanitize_search_term(search_term: str | None) -> str | None:
-    # Maskera söktermer i loggar för att undvika PII-läckage.
-    if not search_term:
-        return search_term
-
-    cleaned = search_term.strip()
-    if not cleaned:
-        return cleaned
-
-    if "@" in cleaned:
-        return mask_email(cleaned)
-
-    digits = "".join(ch for ch in cleaned if ch.isdigit())
-    if len(digits) in {10, 12}:
-        return mask_hash(functions.hash_value(digits))
-
-    if len(cleaned) <= 3:
-        return "***"
-    return f"{cleaned[:1]}***{cleaned[-1:]}"
-
-def _mask_username_for_log(username: str | None) -> str:
-    # Maskera användarnamn före loggning för att undvika persondata i klartext.
-    cleaned = (username or "").strip()
-    if not cleaned:
-        return "<saknas>"
-    if "@" in cleaned:
-        return mask_email(cleaned)
-    if len(cleaned) <= 3:
-        return "***"
-    return f"{cleaned[:2]}***{cleaned[-1]}"
-
-
-
 def _trusted_proxy_hops(raw_value: str | None) -> int:
-    # Tolka TRUSTED_PROXY_COUNT och hantera ogiltiga värden på ett säkert sätt.
-    default_hops = 1
-    if raw_value is None or raw_value.strip() == "":
-        return default_hops
-
-    try:
-        hops = int(raw_value)
-    except ValueError:
-        logger.warning(
-            "Ogiltigt värde för TRUSTED_PROXY_COUNT (%s) – använder standardvärdet 1.",
-            raw_value,
-        )
-        return default_hops
-
-    if hops < 0:
-        logger.warning(
-            "TRUSTED_PROXY_COUNT kan inte vara negativt – proxystödet stängs av.",
-        )
-        return 0
-
-    return hops
+    # Behåll wrappern i app.py så testsviten kan monkeypatcha just denna yta.
+    return bootstrap.trusted_proxy_hops(raw_value)
 
 
 def _configure_proxy_fix(app: Flask) -> None:
-    # Aktivera ProxyFix så att Flask litar på headers från den externa proxyn.
+    # Wrappern använder _trusted_proxy_hops ovan för att bevara tidigare testkrok.
     hops = _trusted_proxy_hops(os.getenv("TRUSTED_PROXY_COUNT"))
-    if hops > 0:
-        app.wsgi_app = ProxyFix(
-            app.wsgi_app,
-            x_for=hops,
-            x_proto=hops,
-            x_host=hops,
-            x_port=hops,
-        )
-        logger.info("Aktiverar ProxyFix med %s betrodda proxyhopp", mask_hash(str(hops)))
-    else:
-        logger.info(
-            "ProxyFix är inaktiverad (TRUSTED_PROXY_COUNT=%s)",
-            os.getenv("TRUSTED_PROXY_COUNT", "0"),
-        )
+    bootstrap.apply_proxy_fix(app, hops)
 
 
 def _configure_timezone() -> str:
-    # Sätt applikationens lokala tidszon för all datum/tid-formattering.
-    timezone_name = os.getenv("APP_TIMEZONE", "Europe/Stockholm").strip() or "Europe/Stockholm"
-    os.environ["TZ"] = timezone_name
-    tzset = getattr(time, "tzset", None)
-    if callable(tzset):
-        tzset()
-    logger.info("Applikationens tidszon är satt till %s.", timezone_name)
-    return timezone_name
-
-
+    return bootstrap.configure_timezone()
 
 
 def _enable_debug_mode(app: Flask) -> None:
-    # Aktivera extra loggning och ev. testdata i debug-läge.
-    if not as_bool(os.getenv("DEV_MODE")):
-        return
+    bootstrap.enable_debug_mode(app)
 
-    stream = logging.StreamHandler()
-    root = logging.getLogger()
-    if not any(isinstance(h, logging.StreamHandler) for h in root.handlers):
-        root.addHandler(stream)
-    # Höj nivån till DEBUG endast i DEV_MODE för felsökning under utveckling.
-    root.setLevel(logging.DEBUG)
 
-    if not any(isinstance(h, logging.StreamHandler) for h in app.logger.handlers):
-        app.logger.addHandler(stream)
-    # Flask-loggaren följer också DEBUG i DEV_MODE för samlad felsökning.
-    app.logger.setLevel(logging.DEBUG)
-
-    # Modulens logger höjs explicit i DEV_MODE för att visa debug-flöden lokalt.
-    logger.setLevel(logging.DEBUG)
-    # Samma gäller paketloggaren i DEV_MODE; produktion styrs av LOG_LEVEL.
-    functions.logger.setLevel(logging.DEBUG)
-    if not any(isinstance(h, logging.StreamHandler) for h in functions.logger.handlers):
-        functions.logger.addHandler(stream)
-
-    functions.logger.debug("Utvecklingsläge är aktiverat")
-    logger.debug("Utvecklingsläge är aktiverat")
 def _is_pytest_running() -> bool:
-    return "PYTEST_CURRENT_TEST" in os.environ or any("pytest" in arg for arg in sys.argv)
+    return bootstrap.is_pytest_running()
 
 
 def _resolve_secret_key() -> str:
-    secret_key = os.getenv("secret_key")
-    if secret_key:
-        return secret_key
-    if _is_pytest_running() and as_bool(os.getenv("DEV_MODE")):
-        logger.warning("secret_key saknas i testmiljön. Genererar temporär nyckel.")
-        return secrets.token_hex(32)
-    error_msg = "KRITISKT: miljövariabeln secret_key måste vara satt och inte tom"
-    logger.critical(error_msg)
-    
-    raise RuntimeError(error_msg)
+    return bootstrap.resolve_secret_key(_is_pytest_running)
 
 
 def create_app() -> Flask:
-    # Create and configure the Flask application.
+    # app.py är nu ett tunt kompositionslager som monterar delmodulerna.
     logger.debug("Applikationen initieras")
     logger.debug("Laddar miljövariabler och initierar databas")
     functions.create_database()
+
     app = Flask(__name__)
     timezone_name = _configure_timezone()
     app.config["APP_TIMEZONE"] = timezone_name
     _configure_proxy_fix(app)
 
-    # Validate secret_key is set
     app.secret_key = _resolve_secret_key()
-
     app.config["MAX_CONTENT_LENGTH"] = UPLOAD_MAX_BYTES
+
     dev_mode = as_bool(os.getenv("DEV_MODE"))
     debug_mode = dev_mode
     app.config["DEV_MODE"] = dev_mode
@@ -568,3685 +124,18 @@ def create_app() -> Flask:
         if debug_mode:
             _enable_debug_mode(app)
 
-    logger.debug("Applikationen skapad och databasen initialiserad")
-    # Email handlers for ERROR and CRITICAL logs are now automatically attached
-    # via configure_root_logging() in functions.logging module.
+    bootstrap.register_app_lifecycle(app)
+    register_public_routes(app)
+    register_supervisor_routes(app)
+    register_user_routes(app)
+    register_admin_routes(app)
+    register_error_handlers(app)
 
     logger.debug("Applikationen är konfigurerad och redo")
     return app
 
 
-
-
 app = create_app()
-
-
-
-@app.before_request
-def _before_first_request():
-    session.permanent = True
-    app.permanent_session_lifetime = timedelta(days=178)
-
-
-@app.before_request
-def _log_request_start() -> None:
-    if _is_pytest_running():
-        return
-
-    # Skip logging for health check endpoint and other non-essential endpoints
-    if request.endpoint in ("health", "robots_txt"):
-        return
-    
-    g.request_start = time.monotonic()
-    g.view_start = g.request_start
-    view_func = app.view_functions.get(request.endpoint) if request.endpoint else None
-    if view_func:
-        g.view_func_name = view_func.__name__
-        logger.debug(
-            "Vy startar: %s args=%s",
-            view_func.__name__,
-            mask_sensitive_data(request.view_args or {}),
-        )
-    client_ip = get_request_ip()
-    logger.debug(
-        "Begäran startad: %s %s (IP=%s, agent=%s)",
-        request.method,
-        request.path,
-        mask_hash(client_ip) if client_ip else "okänd",
-        request.headers.get("User-Agent", "okänd"),
-    )
-    logger.debug(
-        "Begäran headers: %s",
-        mask_headers(dict(request.headers)),
-    )
-    logger.debug(
-        "Begäran query-parametrar: %s",
-        mask_sensitive_data(request.args.to_dict(flat=False)),
-    )
-    if request.is_json:
-        logger.debug(
-            "Begäran JSON-body: %s",
-            mask_sensitive_data(request.get_json(silent=True)),
-        )
-
-
-@app.after_request
-def _log_request_end(response: Response) -> Response:
-    if _is_pytest_running():
-        return response
-
-    # Skip logging for health check endpoint and other non-essential endpoints
-    if request.endpoint in ("health", "robots_txt", "sitemap_xml"):
-        return response
-    
-    start = getattr(g, "request_start", None)
-    duration = time.monotonic() - start if isinstance(start, (int, float)) else 0.0
-    status_code = response.status_code
-    if status_code >= 500:
-        level = logging.ERROR
-    elif status_code >= 400:
-        level = logging.WARNING
-    else:
-        level = logging.INFO
-    logger.log(
-        level,
-        "Begäran slutförd: %s %s -> %s (%.3fs)",
-        request.method,
-        request.path,
-        status_code,
-        duration,
-    )
-    logger.debug(
-        "Svar headers: %s",
-        mask_headers(dict(response.headers)),
-    )
-    if response.is_json:
-        logger.debug(
-            "Svar JSON-body: %s",
-            mask_sensitive_data(response.get_json(silent=True)),
-        )
-    view_start = getattr(g, "view_start", None)
-    view_duration = time.monotonic() - view_start if isinstance(view_start, (int, float)) else 0.0
-    view_func_name = getattr(g, "view_func_name", None)
-    if view_func_name:
-        logger.debug(
-            "Vy avslutad: %s (%.3fs)",
-            view_func_name,
-            view_duration,
-        )
-    return response
-
-
-@app.teardown_request
-def _log_request_exception(exception: BaseException | None) -> None:
-    if exception is not None:
-        logger.error("Undantag under begäran: %s", str(exception))
-
-
-
-
-
-@app.teardown_appcontext
-def _teardown(exception=None):
-    # Called when the app context is being torn down
-    if exception is not None:
-        logger.error("Applikation stängs ner på grund av exception: %s", str(exception))
-        try:
-            critical_events.send_crash_notification(
-                error_message=str(exception), traceback=__import__("traceback").format_exc()
-            )
-        except Exception as e:
-            logger.warning("Kunde inte skicka crash-notifikation: %s", str(e))
-
-
-@app.route("/health")
-def health() -> tuple[dict, int]:
-    # Basic health check endpoint.
-    return {"status": "ok"}, 200
-
-
-def _require_admin() -> str:
-    if not session.get("admin_logged_in"):
-        abort(403)
-    return session.get("admin_username", "okänd")
-
-
-def _require_supervisor() -> tuple[str, str]:
-    if not session.get("supervisor_logged_in"):
-        abort(403)
-    email_hash = session.get("supervisor_email_hash")
-    if not email_hash:
-        abort(403)
-    supervisor_name = session.get("supervisor_name") or functions.get_supervisor_name_by_hash(
-        email_hash
-    )
-    if supervisor_name:
-        session["supervisor_name"] = supervisor_name
-    return email_hash, supervisor_name or "Företagskonto"
-
-
-@app.context_processor
-def inject_flags():
-    # Expose flags indicating debug-läge to Jinja templates.
-    return {
-        "IS_DEV": current_app.debug,
-    }
-
-
-@app.route("/robots.txt")
-def robots_txt():
-    # Serve robots.txt to disallow all crawlers.
-    if app.static_folder is None:
-        abort(404)
-    return send_from_directory(app.static_folder, "robots.txt", mimetype="text/plain")
-
-
-@app.route("/sitemap.xml")
-def sitemap_xml():
-    # Serve sitemap.xml with public URLs only.
-    if app.static_folder is None:
-        abort(404)
-    return send_from_directory(app.static_folder, "sitemap.xml", mimetype="application/xml")
-
-
-@app.route("/.well-known/mta-sts.txt")
-def mta_sts_policy():
-    # Serve MTA-STS policyfilen för den dedikerade mta-sts-domänen.
-    mta_sts_directory = Path(app.root_path) / "deploy" / "mta-sts" / ".well-known"
-    return send_from_directory(mta_sts_directory, "mta-sts.txt", mimetype="text/plain")
-
-
-@app.route("/debug/clear-session", methods=["GET", "POST"])
-def debug_clear_session():
-    if not current_app.config.get("DEV_MODE") or not current_app.debug:
-        abort(404)
-    session.clear()
-    return redirect("/")
-
-
-@app.route("/create_user/<pnr_hash>", methods=["POST", "GET"])
-def create_user(pnr_hash: str):  # type: ignore[no-untyped-def]
-    # Allow a pending user to set a password and activate the account.
-    logger.info("Hanterar create_user för hash %s", pnr_hash)
-    if request.method == "POST":
-        password = request.form.get("password", "").strip()
-        confirm = request.form.get("confirm", "").strip()
-        if password != confirm:
-            return _render_create_supervisor_page(error="Lösenorden måste matcha.", invalid=False)
-        if len(password) < 8:
-            return _render_create_supervisor_page(error="Lösenordet måste vara minst 8 tecken långt.", invalid=False)
-        logger.debug("Skapar användare med hash %s", pnr_hash)
-        if not functions.user_create_user(password, pnr_hash):
-            logger.warning("Kunde inte skapa användare för hash %s", pnr_hash)
-            return _render_create_supervisor_page(
-                error=("Kontot kunde inte aktiveras. Kontrollera att länken är giltig."),
-                invalid=False,
-            )
-        flash(
-            "L\u00f6senordet \u00e4r skapat. Du kan nu logga in p\u00e5 ditt privatkonto.",
-            "success",
-        )
-        return redirect("/login")
-    if functions.check_pending_user_hash(pnr_hash):
-        return render_template(
-            "create_supervisor.html",
-            invalid=False,
-            page_title="Skapa konto",
-            heading="Skapa konto",
-            description=(
-                "Välj ett starkt lösenord för ditt konto. "
-                "Lösenordet måste vara minst åtta tecken långt."
-            ),
-            submit_text="Skapa konto",
-            invalid_message="Länken är ogiltig eller har redan använts.",
-        )
-    logger.warning("Användarhash %s hittades inte under create_user", pnr_hash)
-    abort(404, description="Standardkonto hittades inte")
-
-
-@app.route("/foretagskonto/skapa/<activation_token>", methods=["GET", "POST"])
-def supervisor_create(activation_token: str):
-    pending_email = functions.get_pending_supervisor_email_by_token(activation_token)
-    if request.method == "POST":
-        password = request.form.get("password", "").strip()
-        confirm = request.form.get("confirm", "").strip()
-        if password != confirm:
-            return render_template(
-                "create_supervisor.html",
-                error="Lösenorden måste matcha.",
-                invalid=False,
-            )
-        if not pending_email:
-            return render_template("create_supervisor.html", invalid=True)
-        try:
-            if not functions.supervisor_activate_account(pending_email, password):
-                return render_template(
-                    "create_supervisor.html",
-                    error="Kontot kunde inte aktiveras. Kontrollera att länken är giltig.",
-                    invalid=False,
-                )
-        except ValueError as exc:
-            return render_template(
-                "create_supervisor.html",
-                error=_safe_user_error(
-                    str(exc),
-                    ALLOWED_SUPERVISOR_ACTIVATION_ERRORS,
-                    (
-                        "Kontot kunde inte aktiveras. Kontrollera att länken är giltig "
-                        "och att lösenordet uppfyller kraven."
-                    ),
-                ),
-                invalid=False,
-            )
-        logger.info("Handledarkonto aktiverat för %s", mask_email_reference(pending_email))
-        return redirect(url_for("supervisor_login"))
-
-    if pending_email:
-        return render_template("create_supervisor.html", invalid=False)
-    logger.warning("Handledarens aktiveringslänk hittades inte eller har förbrukats")
-    return render_template("create_supervisor.html", invalid=True)
-
-
-@app.route("/foretagskonto/login", methods=["GET", "POST"])
-def supervisor_login():
-    csrf_token = sec.ensure_csrf_token()
-    if request.method == "POST":
-        if not validate_csrf_token():
-            logger.warning("Ogiltig CSRF-token vid företagskontoinloggning")
-            return (
-                render_template(
-                    "supervisor_login.html",
-                    error="Formuläret är inte längre giltigt. Ladda om sidan och försök igen.",
-                    csrf_token=csrf_token,
-                ),
-                400,
-            )
-        orgnr = request.form.get("orgnr", "").strip()
-        password = request.form.get("password", "").strip()
-        if not orgnr or not password:
-            return render_template(
-                "supervisor_login.html",
-                error="Ogiltiga inloggningsuppgifter.",
-                csrf_token=csrf_token,
-            )
-        try:
-            normalized_orgnr = functions.validate_orgnr(orgnr)
-        except ValueError:
-            logger.warning("Ogiltigt organisationsnummer vid företagskontoinloggning")
-            return render_template(
-                "supervisor_login.html",
-                error="Ogiltiga inloggningsuppgifter.",
-                csrf_token=csrf_token,
-            )
-
-        details = functions.get_supervisor_login_details_for_orgnr(normalized_orgnr)
-        if not details:
-            logger.warning(
-                "Företagskonto saknas för organisationsnummer %s",
-                mask_hash(functions.hash_value(normalized_orgnr)),
-            )
-            return render_template(
-                "supervisor_login.html",
-                error="Ogiltiga inloggningsuppgifter.",
-                csrf_token=csrf_token,
-            )
-
-        normalized_email = details["email"]
-        email_hash = details["email_hash"]
-        try:
-            valid = functions.verify_supervisor_credentials(normalized_email, password)
-        except ValueError:
-            logger.warning(
-                "Ogiltig kontokonfiguration vid företagskontoinloggning för %s",
-                mask_hash(functions.hash_value(normalized_orgnr)),
-            )
-            return render_template(
-                "supervisor_login.html",
-                error="Ogiltiga inloggningsuppgifter.",
-                csrf_token=csrf_token,
-            )
-        if not valid:
-            logger.warning(
-                "Felaktigt lösenord för företagskonto %s",
-                mask_hash(functions.hash_value(normalized_orgnr)),
-            )
-            return render_template(
-                "supervisor_login.html",
-                error="Ogiltiga inloggningsuppgifter.",
-                csrf_token=csrf_token,
-            )
-
-        session["supervisor_logged_in"] = True
-        session["supervisor_email_hash"] = email_hash
-        session["supervisor_orgnr"] = normalized_orgnr
-        supervisor_name = details.get("name") or functions.get_supervisor_name_by_hash(email_hash)
-        if supervisor_name:
-            session["supervisor_name"] = supervisor_name
-        logger.info(
-            "Supervisor %s loggade in för organisationsnummer %s",
-            mask_email_reference(email_hash),
-            mask_hash(functions.hash_value(normalized_orgnr)),
-        )
-        return redirect(url_for("supervisor_dashboard"))
-
-    return render_template("supervisor_login.html", csrf_token=csrf_token)
-
-
-@app.route("/foretagskonto", methods=["GET"])
-def supervisor_dashboard():
-    if not session.get("supervisor_logged_in"):
-        return redirect(url_for("supervisor_login"))
-    email_hash, supervisor_name = _require_supervisor()
-    supervisor_orgnr = (session.get("supervisor_orgnr") or "").strip()
-    csrf_token = sec.ensure_csrf_token()
-    connections = functions.list_supervisor_connections(email_hash)
-    users = []
-    for entry in connections:
-        person_hash = entry["personnummer_hash"]
-        username = (entry.get("username") or "Standardkonto").strip()
-        pdfs = functions.get_user_pdfs(person_hash)
-        for pdf in pdfs:
-            pdf["category_labels"] = labels_for_slugs(pdf.get("categories") or [])
-        users.append(
-            {
-                "personnummer_hash": person_hash,
-                "username": username,
-                "pdfs": pdfs,
-            }
-        )
-
-    company_name = None
-    organization_link_requests = []
-    if supervisor_orgnr:
-        try:
-            overview = functions.get_public_organization_overview(supervisor_orgnr)
-        except ValueError:
-            overview = {}
-        company_name = overview.get("company_name")
-        organization_link_requests = functions.list_pending_organization_link_requests(
-            supervisor_orgnr
-        )
-        status_labels = {
-            "active": "Aktivt privatkonto",
-            "pending": "Inv\u00e4ntar l\u00f6senord",
-            "missing": "Kontot kunde inte hittas",
-        }
-        for entry in organization_link_requests:
-            entry["account_status_label"] = status_labels.get(
-                entry.get("account_status", ""),
-                "Ok\u00e4nd status",
-            )
-
-    return render_template(
-        "supervisor_dashboard.html",
-        company_name=company_name,
-        organization_link_requests=organization_link_requests,
-        supervisor_orgnr=supervisor_orgnr,
-        supervisor_name=supervisor_name,
-        users=users,
-        csrf_token=csrf_token,
-    )
-
-
-def _handle_org_link_request_action(
-    request_id: int,
-    action_func: Callable[[int, str, str], tuple[bool, dict[str, str] | None, str]],
-    email_sender: Callable[[str, str], None],
-    success_message: str,
-    email_failure_message: str,
-    failure_messages: dict[str, str],
-    fallback_message: str,
-    email_log_message: str,
-) -> ResponseReturnValue:
-    email_hash, _ = _require_supervisor()
-    supervisor_orgnr = (session.get("supervisor_orgnr") or "").strip()
-    redirect_target = f"{url_for('supervisor_dashboard')}#organization-link-requests"
-    if not validate_csrf_token():
-        flash(CSRF_EXPIRED_MESSAGE, "error")
-        return redirect(redirect_target)
-    if not supervisor_orgnr:
-        flash("F\u00f6retagskontot saknar organisationsnummer.", "error")
-        return redirect(redirect_target)
-
-    success, request_data, reason = action_func(
-        request_id,
-        email_hash,
-        supervisor_orgnr,
-    )
-    if not success:
-        flash(failure_messages.get(reason, fallback_message), "error")
-        return redirect(redirect_target)
-    if request_data is None:
-        flash(fallback_message, "error")
-        return redirect(redirect_target)
-
-    try:
-        overview = functions.get_public_organization_overview(supervisor_orgnr)
-    except ValueError:
-        company_name = f"organisationsnummer {supervisor_orgnr}"
-    else:
-        company_name = overview.get("company_name") or f"organisationsnummer {supervisor_orgnr}"
-
-    try:
-        email_sender(request_data["user_email"], company_name)
-    except Exception:
-        logger.exception(email_log_message, request_id)
-        flash(email_failure_message, "error")
-    else:
-        flash(success_message, "success")
-
-    return redirect(redirect_target)
-
-
-@app.post("/foretagskonto/organisationskopplingar/<int:request_id>/godkann")
-def supervisor_approve_organization_link_request_route(request_id: int):
-    return _handle_org_link_request_action(
-        request_id,
-        functions.approve_organization_link_request,
-        email_service.send_organization_link_approved_email,
-        "Kopplingen har godk\u00e4nts och privatpersonen har informerats via e-post.",
-        "Kopplingen godk\u00e4ndes men bekr\u00e4ftelsemejlet kunde inte skickas till privatpersonen.",
-        {
-            "missing_request": "F\u00f6rfr\u00e5gan kunde inte hittas.",
-            "missing_user": "Privatkontot finns inte l\u00e4ngre kvar.",
-            "handled_request": "F\u00f6rfr\u00e5gan \u00e4r redan hanterad.",
-        },
-        "Kopplingen kunde inte godk\u00e4nnas.",
-        "Koppling godk\u00e4nd men mejl kunde inte skickas till privatkonto f\u00f6r org-f\u00f6rfr\u00e5gan %s",
-    )
-
-
-@app.post("/foretagskonto/organisationskopplingar/<int:request_id>/avsla")
-def supervisor_reject_organization_link_request_route(request_id: int):
-    return _handle_org_link_request_action(
-        request_id,
-        functions.reject_organization_link_request,
-        email_service.send_organization_link_rejected_email,
-        "F\u00f6rfr\u00e5gan har avslagits och privatpersonen har informerats via e-post.",
-        "F\u00f6rfr\u00e5gan avslogs men bekr\u00e4ftelsemejlet kunde inte skickas till privatpersonen.",
-        {
-            "missing_request": "F\u00f6rfr\u00e5gan kunde inte hittas.",
-            "handled_request": "F\u00f6rfr\u00e5gan \u00e4r redan hanterad.",
-        },
-        "Kopplingen kunde inte avsl\u00e5s.",
-        "Koppling avslogs men mejl kunde inte skickas till privatkonto f\u00f6r org-f\u00f6rfr\u00e5gan %s",
-    )
-
-
-@app.post("/foretagskonto/kopplingsforfragan")
-def supervisor_link_request_route():
-    email_hash, _ = _require_supervisor()
-    if not validate_csrf_token(allow_if_absent=True):
-        flash("Formuläret är inte längre giltigt. Ladda om sidan och försök igen.", "error")
-        return redirect(url_for("supervisor_dashboard"))
-    personnummer = (request.form.get("personnummer") or "").strip()
-    if not personnummer:
-        flash("Ange personnummer för standardkontot.", "error")
-        return redirect(url_for("supervisor_dashboard"))
-    try:
-        normalized = functions.normalize_personnummer(personnummer)
-    except ValueError:
-        flash("Personnumret är ogiltigt.", "error")
-        return redirect(url_for("supervisor_dashboard"))
-
-    success, reason = functions.create_supervisor_link_request(email_hash, normalized)
-    if success:
-        flash("Kopplingsförfrågan har skickats.", "success")
-    elif reason == "missing_user":
-        flash("Standardkontot kunde inte hittas.", "error")
-    elif reason == "already_connected":
-        flash("Standardkontot är redan kopplat till ditt konto.", "error")
-    elif reason == "already_requested":
-        flash("Det finns redan en kopplingsförfrågan till standardkontot.", "error")
-    else:
-        flash("Kopplingsförfrågan kunde inte skickas.", "error")
-    return redirect(url_for("supervisor_dashboard"))
-
-
-@app.route("/foretagskonto/standardkonto/<person_hash>/pdf/<int:pdf_id>")
-def supervisor_download_pdf(person_hash: str, pdf_id: int):
-    email_hash, _ = _require_supervisor()
-    if not functions.supervisor_has_access(email_hash, person_hash):
-        logger.warning(
-            "Handledare %s försökte komma åt pdf %s för %s utan behörighet",
-            email_hash,
-            pdf_id,
-            person_hash,
-        )
-        abort(404)
-    pdf = functions.get_pdf_content(person_hash, pdf_id)
-    if not pdf:
-        abort(404)
-    filename, content = pdf
-    as_attachment = request.args.get("download", "1") != "0"
-    logger.info(
-        "Handledare %s hämtar %s för %s",
-        email_hash,
-        filename,
-        person_hash,
-    )
-    response = make_response(content)
-    response.headers["Content-Type"] = "application/pdf"
-    disposition = "attachment" if as_attachment else "inline"
-    response.headers["Content-Disposition"] = f'{disposition}; filename="{filename}"'
-    return response
-
-
-@app.post("/foretagskonto/dela/<person_hash>/<int:pdf_id>")
-def supervisor_share_pdf_route(person_hash: str, pdf_id: int):
-    email_hash, supervisor_name = _require_supervisor()
-    anchor = request.form.get("anchor", "")
-    redirect_target = url_for("supervisor_dashboard")
-    if anchor:
-        redirect_target += f"#{anchor}"
-
-    if not validate_csrf_token(allow_if_absent=True):
-        flash("Formuläret är inte längre giltigt. Ladda om sidan och försök igen.", "error")
-        return redirect(redirect_target)
-
-    if not functions.supervisor_has_access(email_hash, person_hash):
-        logger.warning(
-            "Handledare %s försökte dela pdf %s för %s utan behörighet",
-            email_hash,
-            pdf_id,
-            person_hash,
-        )
-        flash("Åtgärden kunde inte utföras.", "error")
-        return redirect(redirect_target)
-
-    recipient_email = (request.form.get("recipient_email") or "").strip()
-    if not recipient_email:
-        flash("Ange en e-postadress.", "error")
-        return redirect(redirect_target)
-
-    try:
-        normalized_recipient = email_service.normalize_valid_email(recipient_email)
-    except ValueError:
-        flash("Ogiltig e-postadress.", "error")
-        return redirect(redirect_target)
-
-    pdf = functions.get_pdf_content(person_hash, pdf_id)
-    if not pdf:
-        flash("Intyget kunde inte hittas.", "error")
-        return redirect(redirect_target)
-
-    owner_name = functions.get_username_by_personnummer_hash(person_hash) or "Standardkontot"
-    attachments = [(pdf[0], pdf[1])]
-
-    try:
-        email_service.send_pdf_share_email(
-            normalized_recipient,
-            attachments,
-            supervisor_name,
-            owner_name=owner_name,
-        )
-    except RuntimeError:
-        logger.error(
-            "Misslyckades med att dela pdf %s för %s av handledare %s",
-            pdf_id,
-            person_hash,
-            email_hash,
-        )
-        flash("Ett internt fel inträffade när intyget skulle delas.", "error")
-        return redirect(redirect_target)
-
-    logger.info(
-        "Handledare %s delade pdf %s för %s till %s",
-        email_hash,
-        pdf_id,
-        person_hash,
-        mask_email(normalized_recipient),
-    )
-    flash("Intyget har skickats via e-post.", "success")
-    return redirect(redirect_target)
-
-
-@app.post("/foretagskonto/kopplingar/<person_hash>/ta-bort")
-def supervisor_remove_connection_route(person_hash: str):
-    email_hash, _ = _require_supervisor()
-    anchor = request.form.get("anchor", "")
-    redirect_target = url_for("supervisor_dashboard")
-    if anchor:
-        redirect_target += f"#{anchor}"
-
-    if not validate_csrf_token(allow_if_absent=True):
-        flash("Formuläret är inte längre giltigt. Ladda om sidan och försök igen.", "error")
-        return redirect(redirect_target)
-
-    if functions.supervisor_remove_connection(email_hash, person_hash):
-        logger.info(
-            "Handledare %s tog bort åtkomst till %s",
-            mask_email_reference(email_hash),
-            person_hash,
-        )
-        flash("Kopplingen har tagits bort.", "success")
-    else:
-        flash("Kopplingen kunde inte tas bort.", "error")
-    return redirect(redirect_target)
-
-
-@app.route("/aterstall-losenord/<token>", methods=["GET", "POST"])
-def password_reset(token: str):
-    info = functions.get_password_reset(token)
-    if not info or info.get("used_at") is not None:
-        return render_template("password_reset.html", invalid=True)
-
-    if request.method == "POST":
-        password = request.form.get("password", "").strip()
-        confirm = request.form.get("confirm", "").strip()
-        if not password or password != confirm:
-            return render_template(
-                "password_reset.html",
-                invalid=False,
-                error="Lösenorden måste fyllas i och matcha.",
-            )
-        if len(password) < 8:
-            return render_template(
-                "password_reset.html",
-                invalid=False,
-                error="Lösenordet måste vara minst 8 tecken.",
-            )
-        if not functions.reset_password_with_token(token, password):
-            return render_template("password_reset.html", invalid=True)
-        return redirect("/login")
-
-    return render_template("password_reset.html", invalid=False)
-
-
-@app.route("/foretagskonto/aterstall-losenord/<token>", methods=["GET", "POST"])
-def supervisor_password_reset(token: str):
-    info = functions.get_supervisor_password_reset(token)
-    if not info or info.get("used_at") is not None:
-        return render_template("supervisor_password_reset.html", invalid=True)
-
-    if request.method == "POST":
-        password = request.form.get("password", "").strip()
-        confirm = request.form.get("confirm", "").strip()
-        if not password or password != confirm:
-            return render_template(
-                "supervisor_password_reset.html",
-                invalid=False,
-                error="Lösenorden måste fyllas i och matcha.",
-            )
-        if len(password) < 8:
-            return render_template(
-                "supervisor_password_reset.html",
-                invalid=False,
-                error="Lösenordet måste vara minst 8 tecken.",
-            )
-        if not functions.reset_supervisor_password_with_token(token, password):
-            return render_template("supervisor_password_reset.html", invalid=True)
-        return redirect(url_for("supervisor_login"))
-
-    return render_template("supervisor_password_reset.html", invalid=False)
-
-
-@app.route("/", methods=["GET"])
-def home():
-    # Render the landing page.
-    logger.debug("Renderar startsida")
-    return render_template("index.html")
-
-
-@app.route("/ansok", methods=["GET"])
-def apply_account():
-    """Visa val för ansökan om konto."""
-
-    return render_template("apply.html")
-
-
-def _flag_application_field_error(message: str, field_errors: dict[str, bool]) -> None:
-    # Markera specifika fält som ska flaggas i formuläret utifrån felmeddelanden.
-    lowered = message.lower()
-    if "namn" in lowered:
-        field_errors["name"] = True
-    if "e-post" in lowered or "email" in lowered:
-        field_errors["email"] = True
-    if "personnummer" in lowered:
-        field_errors["personnummer"] = True
-    if "organisationsnummer" in lowered:
-        field_errors["orgnr"] = True
-    if "företagsnamn" in lowered:
-        field_errors["company_name"] = True
-    if "fakturaadress" in lowered:
-        field_errors["invoice_address"] = True
-    if "kontaktperson" in lowered:
-        field_errors["invoice_contact"] = True
-    if "märkning" in lowered:
-        field_errors["invoice_reference"] = True
-
-
-def _handle_standard_account_registration():
-    form_errors: list[str] = []
-    field_errors = {
-        "name": False,
-        "email": False,
-        "personnummer": False,
-        "terms_confirmed": False,
-    }
-    form_data = {
-        "name": "",
-        "email": "",
-        "personnummer": "",
-        "terms_confirmed": "",
-    }
-    status_code = 200
-
-    if request.method == "POST":
-        for key in form_data:
-            form_data[key] = (request.form.get(key, "") or "").strip()
-        if not validate_csrf_token():
-            form_errors.append(CSRF_EXPIRED_MESSAGE)
-        else:
-            client_ip = get_request_ip()
-            if not register_public_submission(client_ip):
-                status_code = 429
-                form_errors.append(TOO_MANY_ATTEMPTS_MESSAGE)
-            else:
-                if not as_bool(form_data.get("terms_confirmed")):
-                    field_errors["terms_confirmed"] = True
-                    form_errors.append(
-                        "Du m\u00e5ste intyga att du har l\u00e4st och f\u00f6rst\u00e5tt villkoren och den juridiska informationen innan du skapar kontot."
-                    )
-                if not form_errors:
-                    class _ActivationEmailDeliveryError(RuntimeError):
-                        pass
-
-                    try:
-                        def _send_creation_email_before_commit(
-                            registration_result: dict[str, str],
-                        ) -> None:
-                            creation_link = url_for(
-                                "create_user",
-                                pnr_hash=registration_result["personnummer_hash"],
-                                _external=True,
-                            )
-                            try:
-                                email_service.send_creation_email(
-                                    registration_result["email"],
-                                    creation_link,
-                                )
-                            except Exception as exc:
-                                raise _ActivationEmailDeliveryError from exc
-
-                        functions.register_standard_account(
-                            form_data["name"],
-                            form_data["email"],
-                            form_data["personnummer"],
-                            before_commit=_send_creation_email_before_commit,
-                        )
-                        logger.info(
-                            "Nytt privatkonto registrerat f\u00f6r %s",
-                            mask_hash(functions.hash_value(form_data["email"].lower())),
-                        )
-                        return redirect(url_for("standard_account_registered"))
-                    except ValueError as exc:
-                        message = str(exc)
-                        form_errors.append(message)
-                        _flag_application_field_error(message, field_errors)
-                    except _ActivationEmailDeliveryError:
-                        logger.exception(
-                            "Privatkonto kunde inte skapas eftersom aktiveringsmejlet inte gick att skicka"
-                        )
-                        form_errors.append(
-                            "Det gick inte att skicka aktiveringsmejlet just nu. Försök igen senare."
-                        )
-                    except Exception:
-                        logger.exception("Kunde inte skapa privatkonto")
-                        form_errors.append(
-                            "Det gick inte att skapa kontot just nu. F\u00f6rs\u00f6k igen senare."
-                        )
-
-    csrf_token = sec.ensure_csrf_token()
-    return (
-        render_template(
-            "apply_standardkonto.html",
-            csrf_token=csrf_token,
-            form_data=form_data,
-            form_errors=form_errors,
-            field_errors=field_errors,
-        ),
-        status_code,
-    )
-
-
-@app.route("/ansok/standardkonto", methods=["GET", "POST"])
-def apply_standardkonto():
-    """Visa och hantera ansökan för standardkonto."""
-
-    return _handle_standard_account_registration()
-
-
-@app.route("/ansok/standardkonto/klart", methods=["GET"])
-def standard_account_registered():
-    """Visa bekräftelse efter skapad standardkontoansökan."""
-
-    return render_template("standard_account_registered.html")
-
-
-@app.route("/ansok/foretagskonto", methods=["GET", "POST"])
-def apply_foretagskonto():
-    """Visa och hantera ansökan för företagskonto."""
-
-    account_type = "foretagskonto"
-    form_errors: list[str] = []
-    base_field_errors = {
-        "name": False,
-        "email": False,
-        "orgnr": False,
-        "company_name": False,
-        "invoice_address": False,
-        "invoice_contact": False,
-        "invoice_reference": False,
-        "comment": False,
-        "terms_confirmed": False,
-    }
-    field_errors = dict(base_field_errors)
-    status_code = 200
-
-    base_form_data = {
-        "name": "",
-        "email": "",
-        "orgnr": "",
-        "comment": "",
-        "terms_confirmed": "",
-        "company_name": "",
-        "invoice_address": "",
-        "invoice_contact": "",
-        "invoice_reference": "",
-    }
-
-    form_data = dict(base_form_data)
-
-    if request.method == "POST":
-        for key in form_data:
-            form_data[key] = (request.form.get(key, "") or "").strip()
-        if not validate_csrf_token():
-            form_errors.append("Formuläret är inte längre giltigt. Ladda om sidan och försök igen.")
-        else:
-            client_ip = get_request_ip()
-            if not register_public_submission(client_ip):
-                status_code = 429
-                form_errors.append("Du har gjort för många försök. Vänta en stund och prova igen.")
-            else:
-                if not as_bool(form_data.get("terms_confirmed")):
-                    field_errors["terms_confirmed"] = True
-                    form_errors.append(
-                        "Du måste intyga att du har läst och förstått villkoren och den juridiska informationen innan du skickar ansökan."
-                    )
-                if not form_errors:
-                    try:
-                        request_id = functions.create_application_request(
-                            account_type,
-                            form_data["name"],
-                            form_data["email"],
-                            form_data["orgnr"],
-                            form_data.get("company_name"),
-                            form_data.get("comment"),
-                            form_data.get("invoice_address"),
-                            form_data.get("invoice_contact"),
-                            form_data.get("invoice_reference"),
-                        )
-                        logger.info(
-                            "Ny ansökan %s mottagen från %s",
-                            request_id,
-                            mask_hash(functions.hash_value(form_data["email"].lower())),
-                        )
-                        try:
-                            email_service.new_application_email_to_support(
-                                account_type=account_type
-                            )
-                        except Exception:  # pragma: no cover - e-postfel ska inte stoppa ansökan
-                            logger.exception(
-                                "Ansökan sparades men supportmejl kunde inte skickas"
-                            )
-                    except ValueError as exc:
-                        message = str(exc)
-                        form_errors.append(message)
-                        _flag_application_field_error(message, field_errors)
-                    except Exception:  # pragma: no cover - defensiv loggning
-                        logger.exception("Kunde inte spara ansökan")
-                        form_errors.append(
-                            "Det gick inte att skicka ansökan just nu. Försök igen senare."
-                        )
-                    else:
-                        # Make the client-facing confirmation explicit about which account type was submitted
-                        display_type = (
-                            "företagskonto" if account_type == "foretagskonto" else "standardkonto"
-                        )
-                        flash(
-                            f"Din ansökan om {display_type} har skickats. Tack! Vi hör av oss så snart vi granskat ansökan.",
-                            "success",
-                        )
-                        return redirect(url_for("application_submitted", account_type=account_type))
-
-    csrf_token = sec.ensure_csrf_token()
-
-    return (
-        render_template(
-            "apply_foretagskonto.html",
-            csrf_token=csrf_token,
-            form_data=form_data,
-            form_errors=form_errors,
-            field_errors=field_errors,
-        ),
-        status_code,
-    )
-
-
-@app.route("/villkor", methods=["GET"])
-def terms_of_service():
-    """Visa sidan med villkor."""
-
-    return render_template("terms_of_service.html")
-
-
-@app.route("/ansok/klart", methods=["GET"])
-def application_submitted():
-    """Visa bekräftelse och nästa steg efter inskickad ansökan."""
-
-    raw_type = request.args.get("account_type", "").strip().lower()
-    account_type = "företagskonto" if raw_type == "foretagskonto" else "standardkonto"
-    return render_template("application_submitted.html", account_type=account_type)
-
-
-@app.route("/organisationer", methods=["GET"])
-def public_organization_search():
-    search_value = (request.args.get("orgnr") or "").strip()
-    result = None
-    form_error = None
-
-    if search_value:
-        try:
-            result = functions.get_public_organization_overview(search_value)
-        except ValueError:
-            form_error = "Kontrollera organisationsnumret och f\u00f6rs\u00f6k igen."
-
-    return render_template(
-        "public_organization_search.html",
-        form_data={"orgnr": search_value},
-        form_error=form_error,
-        result=result,
-    )
-
-
-@app.route("/pris", methods=["GET"])
-def pricing():
-    """Visa prislistan."""
-
-    return render_template("pris.html")
-
-
-@app.route("/login", methods=["GET", "POST"])
-def login():
-    # Authenticate users using personnummer and password.
-    csrf_token = sec.ensure_csrf_token()
-    if request.method == "POST":
-        if not validate_csrf_token():
-            logger.warning("Ogiltig CSRF-token vid användarinloggning")
-            return (
-                render_template(
-                    "user_login.html",
-                    error="Formuläret är inte längre giltigt. Ladda om sidan och försök igen.",
-                    csrf_token=csrf_token,
-                ),
-                400,
-            )
-        raw_personnummer = request.form["personnummer"]
-        try:
-            personnummer = functions.normalize_personnummer(raw_personnummer)
-        except ValueError:
-            logger.warning("Ogiltigt personnummer angivet vid inloggning")
-            return (
-                render_template(
-                    "user_login.html",
-                    error="Ogiltiga inloggningsuppgifter",
-                    csrf_token=csrf_token,
-                ),
-                401,
-            )
-
-        if personnummer == "" or not personnummer.isnumeric():
-            logger.warning("Ogiltigt normaliserat personnummer vid inloggning")
-            return (
-                render_template(
-                    "user_login.html",
-                    error="Ogiltiga inloggningsuppgifter",
-                    csrf_token=csrf_token,
-                ),
-                401,
-            )
-        password = request.form["password"]
-        personnummer_hash = functions.hash_value(personnummer)
-        if password == "":
-            logger.warning("Tomt lösenord angavs för %s", mask_hash(personnummer_hash))
-            return (
-                render_template(
-                    "user_login.html",
-                    error="Ogiltiga inloggningsuppgifter",
-                    csrf_token=csrf_token,
-                ),
-                401,
-            )
-        logger.debug("Inloggningsförsök för %s", mask_hash(personnummer_hash))
-        if functions.check_personnummer_password(personnummer, password):
-            session["user_logged_in"] = True
-            session["personnummer"] = personnummer_hash
-            session["personnummer_raw"] = personnummer
-            session["username"] = functions.get_username_by_personnummer_hash(personnummer_hash)
-            logger.info("Användare %s loggade in", mask_hash(personnummer_hash))
-            return redirect("/dashboard")
-
-        error_message = "Ogiltiga inloggningsuppgifter"
-        if functions.check_pending_user(personnummer):
-            logger.warning(
-                "Inloggning nekades för ej verifierat konto %s",
-                mask_hash(personnummer_hash),
-            )
-            error_message = (
-                "Du behöver verifiera din e-postadress via länken i mejlet innan du kan logga in."
-            )
-        else:
-            logger.warning("Ogiltig inloggning för %s", mask_hash(personnummer_hash))
-        return (
-            render_template(
-                "user_login.html",
-                error=error_message,
-                csrf_token=csrf_token,
-            ),
-            401,
-        )
-    logger.debug("Renderar inloggningssida")
-    return render_template("user_login.html", csrf_token=csrf_token)
-
-
-@app.route("/dashboard", methods=["GET"])
-def dashboard():
-    # Visa alla PDF:er för den inloggade användaren.
-    if not session.get("user_logged_in"):
-        logger.debug("Oautentiserad åtkomst till dashboard")
-        return redirect("/login")
-    pnr_hash = session.get("personnummer")
-    user_name = session.get("username")
-    if not user_name and pnr_hash:
-        user_name = functions.get_username_by_personnummer_hash(pnr_hash)
-        if user_name:
-            session["username"] = user_name
-    if not pnr_hash:
-        return redirect("/login")
-    pdfs = functions.get_user_pdfs(pnr_hash)
-    for pdf in pdfs:
-        pdf["category_labels"] = labels_for_slugs(pdf.get("categories", []))
-        pdf["editable_name"] = _editable_pdf_name(pdf.get("filename", ""))
-    grouped_pdfs = []
-    groups_by_slug = {}
-    for slug, label in COURSE_CATEGORIES:
-        group = {"slug": slug, "label": label, "pdfs": []}
-        grouped_pdfs.append(group)
-        groups_by_slug[slug] = group
-    uncategorized_group = {
-        "slug": "okategoriserade",
-        "label": "Okategoriserade intyg",
-        "pdfs": [],
-    }
-    for pdf in pdfs:
-        categories = pdf.get("categories") or []
-        matched = False
-        for slug in categories:
-            group = groups_by_slug.get(slug)
-            if group is not None:
-                group["pdfs"].append(pdf)
-                matched = True
-        if not matched:
-            uncategorized_group["pdfs"].append(pdf)
-    visible_groups = [group for group in grouped_pdfs if group["pdfs"]]
-    if uncategorized_group["pdfs"]:
-        visible_groups.append(uncategorized_group)
-    category_summary = [
-        {
-            "slug": slug,
-            "label": label,
-            "count": len(groups_by_slug[slug]["pdfs"]),
-        }
-        for slug, label in COURSE_CATEGORIES
-    ]
-    if uncategorized_group["pdfs"]:
-        category_summary.append(
-            {
-                "slug": uncategorized_group["slug"],
-                "label": uncategorized_group["label"],
-                "count": len(uncategorized_group["pdfs"]),
-            }
-        )
-    pending_link_requests = functions.list_user_link_requests(pnr_hash)
-    supervisor_connections = functions.list_user_supervisor_connections(pnr_hash)
-    logger.debug("Dashboard för %s visar %d pdfer", pnr_hash, len(pdfs))
-    user_name = _format_display_name(user_name)
-    csrf_token = sec.ensure_csrf_token()
-    return render_template(
-        "dashboard.html",
-        pdfs=pdfs,
-        course_categories=COURSE_CATEGORIES,
-        course_category_groups=COURSE_CATEGORY_GROUPS,
-        category_summary=category_summary,
-        grouped_pdfs=visible_groups,
-        user_name=user_name,
-        pending_link_requests=pending_link_requests,
-        supervisor_connections=supervisor_connections,
-        csrf_token=csrf_token,
-        today_iso=date.today().isoformat(),
-    )
-
-
-@app.post("/dashboard/ladda-upp")
-def user_upload_pdf_route():
-    if not session.get("user_logged_in"):
-        return redirect("/login")
-
-    if not validate_csrf_token():
-        flash("Formuläret är inte längre giltigt. Ladda om sidan och försök igen.", "error")
-        return redirect("/dashboard")
-
-    personnummer = session.get("personnummer_raw")
-    if not personnummer:
-        flash("Kunde inte identifiera användaren. Logga in igen.", "error")
-        return redirect("/dashboard")
-
-    if not validate_csrf_token(allow_if_absent=True):
-        flash("Formuläret är inte längre giltigt. Ladda om sidan och försök igen.", "error")
-        return redirect("/dashboard")
-
-    uploaded_file = request.files.get("certificate")
-    category = request.form.get("category", "")
-    note = (request.form.get("note", "") or "").strip()
-    expiry_mode = request.form.get("expiry_mode", "none")
-    expiry_date_raw = request.form.get("expiry_date", "")
-    expiry_months_raw = request.form.get("expiry_months", "")
-    expiry_years_raw = request.form.get("expiry_years", "")
-
-    if not uploaded_file or uploaded_file.filename == "":
-        flash("Ingen fil vald.", "error")
-        return redirect("/dashboard")
-
-    if not category:
-        flash("Välj en kurskategori.", "error")
-        return redirect("/dashboard")
-
-    if len(note) > 300:
-        flash("Anteckningen får vara högst 300 tecken.", "error")
-        return redirect("/dashboard")
-
-    try:
-        expires_on = _resolve_certificate_expiry(
-            expiry_mode,
-            expiry_date_raw,
-            expiry_months_raw,
-            expiry_years_raw,
-        )
-    except ValueError as exc:
-        flash(str(exc), "error")
-        return redirect("/dashboard")
-
-    try:
-        pdf.save_pdf_for_user(
-            personnummer,
-            uploaded_file,
-            [category],
-            note=note,
-            expires_on=expires_on,
-            logger=logger,
-        )
-    except ValueError as exc:
-        flash(
-            _safe_user_error(
-                str(exc),
-                ALLOWED_PDF_UPLOAD_ERRORS,
-                "Filen kunde inte laddas upp. Kontrollera filformatet och försök igen.",
-            ),
-            "error",
-        )
-        return redirect("/dashboard")
-    except Exception:
-        logger.exception("Kunde inte spara intyg för användare")
-        flash("Ett fel inträffade när intyget skulle sparas.", "error")
-        return redirect("/dashboard")
-
-    flash("Intyget har laddats upp och sparats som PDF.", "success")
-    return redirect("/dashboard")
-
-
-@app.post("/dashboard/kopplingsforfragan/<supervisor_hash>/godkann")
-def user_accept_link_request_route(supervisor_hash: str):
-    if not session.get("user_logged_in"):
-        return redirect("/login")
-    if not validate_csrf_token(allow_if_absent=True):
-        flash("Formuläret är inte längre giltigt. Ladda om sidan och försök igen.", "error")
-        return redirect("/dashboard")
-    personnummer_hash = session.get("personnummer")
-    if not personnummer_hash:
-        flash("Inte inloggad.", "error")
-        return redirect("/login")
-    if functions.user_accept_link_request(personnummer_hash, supervisor_hash):
-        flash("Kopplingen är nu aktiv.", "success")
-    else:
-        flash("Kopplingsförfrågan kunde inte godkännas.", "error")
-    return redirect("/dashboard")
-
-
-@app.post("/dashboard/kopplingsforfragan/<supervisor_hash>/avsla")
-def user_reject_link_request_route(supervisor_hash: str):
-    if not session.get("user_logged_in"):
-        return redirect("/login")
-    if not validate_csrf_token(allow_if_absent=True):
-        flash("Formuläret är inte längre giltigt. Ladda om sidan och försök igen.", "error")
-        return redirect("/dashboard")
-    personnummer_hash = session.get("personnummer")
-    if not personnummer_hash:
-        flash("Inte inloggad.", "error")
-        return redirect("/login")
-    if functions.user_reject_link_request(personnummer_hash, supervisor_hash):
-        flash("Kopplingsförfrågan har avböjts.", "success")
-    else:
-        flash("Kopplingsförfrågan kunde inte avböjas.", "error")
-    return redirect("/dashboard")
-
-
-@app.post("/dashboard/kopplingar/<supervisor_hash>/ta-bort")
-def user_remove_supervisor_connection_route(supervisor_hash: str):
-    if not session.get("user_logged_in"):
-        return redirect("/login")
-    if not validate_csrf_token(allow_if_absent=True):
-        flash("Formuläret är inte längre giltigt. Ladda om sidan och försök igen.", "error")
-        return redirect("/dashboard")
-    personnummer_hash = session.get("personnummer")
-    if not personnummer_hash:
-        flash("Inte inloggad.", "error")
-        return redirect("/login")
-    if functions.user_remove_supervisor_connection(personnummer_hash, supervisor_hash):
-        flash("Kopplingen har tagits bort.", "success")
-    else:
-        flash("Kopplingen kunde inte tas bort.", "error")
-    return redirect("/dashboard")
-
-
-@app.post("/dashboard/kopplingar/id/<int:connection_id>/ta-bort")
-def user_remove_supervisor_connection_by_id_route(connection_id: int):
-    if not session.get("user_logged_in"):
-        return redirect("/login")
-    if not validate_csrf_token(allow_if_absent=True):
-        flash("Formuläret är inte längre giltigt. Ladda om sidan och försök igen.", "error")
-        return redirect("/dashboard")
-    personnummer_hash = session.get("personnummer")
-    if not personnummer_hash:
-        flash("Inte inloggad.", "error")
-        return redirect("/login")
-    if functions.user_remove_supervisor_connection_by_id(personnummer_hash, connection_id):
-        flash("Kopplingen har tagits bort.", "success")
-    else:
-        flash("Kopplingen kunde inte tas bort.", "error")
-    return redirect("/dashboard")
-
-
-@app.post("/dashboard/intyg/<int:pdf_id>/ta-bort")
-def user_delete_pdf_route(pdf_id: int):
-    if not session.get("user_logged_in"):
-        return redirect("/login")
-
-    if not validate_csrf_token(allow_if_absent=True):
-        flash("Formuläret är inte längre giltigt. Ladda om sidan och försök igen.", "error")
-        return redirect("/dashboard")
-
-    personnummer = session.get("personnummer_raw")
-    if not personnummer:
-        flash("Kunde inte identifiera användaren. Logga in igen.", "error")
-        return redirect("/dashboard")
-
-    try:
-        if functions.delete_user_pdf(personnummer, pdf_id):
-            flash("Intyget har tagits bort.", "success")
-        else:
-            flash("Intyget kunde inte tas bort.", "error")
-    except Exception:
-        logger.error("Kunde inte ta bort intyg %s för användare", pdf_id)
-        flash("Ett fel inträffade när intyget skulle tas bort.", "error")
-
-    return redirect("/dashboard")
-
-
-@app.post("/dashboard/intyg/<int:pdf_id>/uppdatera")
-def user_update_pdf_route(pdf_id: int):
-    if not session.get("user_logged_in"):
-        return jsonify({"fel": "Du måste vara inloggad för att uppdatera intyg."}), 401
-
-    if not validate_csrf_token():
-        return jsonify({"fel": CSRF_EXPIRED_MESSAGE}), 400
-
-    personnummer = session.get("personnummer_raw")
-    if not personnummer:
-        return jsonify({"fel": "Kunde inte identifiera användaren. Logga in igen."}), 400
-    personnummer_hash = functions.hash_value(personnummer)
-
-    payload = request.get_json(silent=True) or {}
-    if not isinstance(payload, dict):
-        return jsonify({"fel": "Ogiltig begäran."}), 400
-
-    try:
-        filename_raw = _coerce_text_payload_value(
-            payload.get("filename", ""),
-            "Intygsnamnet",
-        ).strip()
-        note = _coerce_text_payload_value(payload.get("note", ""), "Anteckningen").strip()
-        expiry_mode = _coerce_text_payload_value(
-            payload.get("expiry_mode", "none"),
-            "Val för utgångsdatum",
-            default="none",
-        ).strip()
-        expiry_date_raw = _coerce_text_payload_value(
-            payload.get("expiry_date", ""),
-            "Utgångsdatum",
-        ).strip()
-        expiry_months_raw = _coerce_text_payload_value(
-            payload.get("expiry_months", ""),
-            "Antal månader",
-        ).strip()
-        expiry_years_raw = _coerce_text_payload_value(
-            payload.get("expiry_years", ""),
-            "Antal år",
-        ).strip()
-    except ValueError as exc:
-        current_app.logger.info("PDF update payload validation failed: %s", exc)
-        return (
-            jsonify(
-                {
-                    "fel": _safe_user_error(
-                        str(exc),
-                        ALLOWED_PDF_METADATA_UPDATE_ERRORS,
-                        "Ogiltig begäran.",
-                    )
-                }
-            ),
-            400,
-        )
-
-    if len(note) > 300:
-        return jsonify({"fel": "Anteckningen får vara högst 300 tecken."}), 400
-
-    current_pdf = functions.get_pdf_metadata(personnummer_hash, pdf_id)
-    current_expires_on = current_pdf.get("expires_on") if current_pdf else None
-
-    try:
-        filename = pdf.build_editable_pdf_filename(filename_raw)
-        expires_on = _resolve_certificate_expiry(
-            expiry_mode,
-            expiry_date_raw,
-            expiry_months_raw,
-            expiry_years_raw,
-            current_expires_on=current_expires_on,
-        )
-    except ValueError as exc:
-        current_app.logger.info("PDF metadata update validation failed: %s", exc)
-        return jsonify({"fel": str(exc)}), 400
-
-    try:
-        updated = functions.update_user_pdf_metadata(
-            personnummer,
-            pdf_id,
-            filename,
-            note,
-            expires_on,
-        )
-    except Exception:
-        logger.exception("Kunde inte uppdatera intyg %s för användare", pdf_id)
-        return jsonify({"fel": "Ett fel inträffade när intyget skulle uppdateras."}), 500
-
-    if not updated:
-        return jsonify({"fel": "Intyget kunde inte hittas."}), 404
-
-    return (
-        jsonify(
-            {
-                "meddelande": "Intyget har uppdaterats.",
-                "data": {
-                    "id": pdf_id,
-                    "filename": filename,
-                    "note": note,
-                    "expires_on": (
-                        expires_on.isoformat()
-                        if expires_on and hasattr(expires_on, "isoformat")
-                        else None
-                    ),
-                },
-            }
-        ),
-        200,
-    )
-
-@app.route("/dashboard/upload", methods=["GET"])
-def user_upload_page():
-    if not session.get("user_logged_in"):
-        logger.debug("Oautentiserad åtkomst till uppladdningssidan")
-        return redirect("/login")
-
-    personnummer_hash = session.get("personnummer")
-    if not personnummer_hash:
-        return redirect("/login")
-
-    user_name = session.get("username")
-    if not user_name:
-        user_name = functions.get_username_by_personnummer_hash(personnummer_hash)
-        if user_name:
-            session["username"] = user_name
-
-    csrf_token = sec.ensure_csrf_token()
-    certificate_count = functions.count_user_pdfs(personnummer_hash)
-    return render_template(
-        "upload_intyg.html",
-        course_categories=COURSE_CATEGORIES,
-        course_category_groups=COURSE_CATEGORY_GROUPS,
-        csrf_token=csrf_token,
-        user_name=_format_display_name(user_name),
-        certificate_count=certificate_count,
-        today_iso=date.today().isoformat(),
-        max_upload_mb=UPLOAD_MAX_MB,
-        max_upload_bytes=UPLOAD_MAX_BYTES,
-    )
-
-
-@app.route("/my_pdfs/<int:pdf_id>")
-def download_pdf(pdf_id: int):
-    # Serve a stored PDF for the logged-in user from the database.
-    if not session.get("user_logged_in"):
-        logger.debug("Oautentiserat nedladdningsförsök för %s", pdf_id)
-        return redirect("/login")
-    pnr_hash = session.get("personnummer")
-    if not pnr_hash:
-        return redirect("/login")
-    as_attachment = request.args.get("download", "1") != "0"
-    pdf = functions.get_pdf_content(pnr_hash, pdf_id)
-    if not pdf:
-        logger.warning("PDF %s hittades inte för användare %s", pdf_id, pnr_hash)
-        abort(404)
-    filename, content = pdf
-    logger.info("Användare %s hämtar %s (as_attachment=%s)", pnr_hash, filename, as_attachment)
-    response = make_response(content)
-    response.headers["Content-Type"] = "application/pdf"
-    disposition = "attachment" if as_attachment else "inline"
-    response.headers["Content-Disposition"] = f'{disposition}; filename="{filename}"'
-    return response
-
-
-@app.route("/share_pdf", methods=["POST"])
-def share_pdf() -> tuple[Response, int]:  # pragma: no cover
-    # Share a PDF with a recipient via e-post.
-    if not session.get("user_logged_in"):
-        logger.debug("Oautentiserat delningsförsök")
-        return jsonify({"fel": "Du måste vara inloggad för att dela intyg."}), 401
-
-    payload = request.get_json(silent=True) or request.form
-    if not payload:
-        logger.error("Tom payload i share_pdf: %r", payload)
-        return jsonify({"fel": "Ogiltig begäran."}), 400
-
-    pdf_ids_raw = payload.get("pdf_ids") if hasattr(payload, "get") else None
-    recipient_email = (
-        payload.get("recipient_email", "") if hasattr(payload, "get") else ""
-    ).strip()
-
-    if pdf_ids_raw is None and hasattr(payload, "get"):
-        pdf_id_raw = payload.get("pdf_id")
-        if pdf_id_raw is not None:
-            pdf_ids_raw = [pdf_id_raw]
-
-    if pdf_ids_raw is None:
-        logger.debug("Inga pdf_ids angavs i share_pdf: %r", payload)
-        return jsonify({"fel": "Ogiltigt intyg angivet."}), 400
-
-    if isinstance(pdf_ids_raw, (str, bytes)):
-        candidate_ids = [pdf_ids_raw]
-    elif isinstance(pdf_ids_raw, (list, tuple, set)):
-        candidate_ids = list(pdf_ids_raw)
-    else:
-        candidate_ids = [pdf_ids_raw]
-
-    pdf_ids: list[int] = []
-    seen_ids: set[int] = set()
-    for raw_id in candidate_ids:
-        try:
-            pdf_id = int(raw_id)
-        except (TypeError, ValueError):
-            logger.warning("Ogiltigt pdf_id angavs för delning: %r", raw_id)
-            return jsonify({"fel": "Ogiltigt intyg angivet."}), 400
-        if pdf_id in seen_ids:
-            continue
-        seen_ids.add(pdf_id)
-        pdf_ids.append(pdf_id)
-
-    if not pdf_ids:
-        logger.debug("Tomma pdf_ids efter bearbetning i share_pdf: %r", payload)
-        return jsonify({"fel": "Ogiltigt intyg angivet."}), 400
-
-    if not recipient_email:
-        logger.debug("Tom recipient_email i share_pdf: %r", payload)
-        return jsonify({"fel": "Ange en e-postadress."}), 400
-
-    pnr_hash = session.get("personnummer")
-    if not pnr_hash:
-        logger.error("Delningsbegäran saknar personnummer i session. Nycklar: %s", sorted(session.keys()))
-        return jsonify({"fel": "Saknar användaruppgifter."}), 400
-
-    attachments: list[tuple[str, bytes]] = []
-
-    for pdf_id in pdf_ids:
-        pdf = functions.get_pdf_content(pnr_hash, pdf_id)
-        if not pdf:
-            logger.debug("PDF %s hittades inte för användare %s vid delning", pdf_id, pnr_hash)
-            return jsonify({"fel": "Intyget kunde inte hittas."}), 404
-        filename, content = pdf
-        attachments.append((filename, content))
-
-    sender_name = session.get("username")
-    if not sender_name:
-        sender_name = functions.get_username_by_personnummer_hash(pnr_hash)
-        if sender_name:
-            session["username"] = sender_name
-
-    sender_display = (sender_name or "").strip() or "Ett standardkonto"
-
-    try:
-        normalized_recipient = email_service.normalize_valid_email(recipient_email)
-    except ValueError:
-        logger.debug("Ogiltig recipient_email i share_pdf: %s", mask_email(recipient_email))
-        return jsonify({"fel": "Ogiltig e-postadress."}), 400
-
-    if normalized_recipient != recipient_email:
-        logger.debug(
-            "Normaliserade mottagaradress för delning från %r till %s",
-            mask_email(recipient_email),
-            mask_email(normalized_recipient),
-        )
-
-    try:
-        email_service.send_pdf_share_email(
-            normalized_recipient,
-            attachments,
-            sender_display,
-        )
-    except RuntimeError as exc:
-        logger.error(
-            "Misslyckades med att dela pdf %s från %s till %s. Fel: %s",
-            pdf_ids,
-            pnr_hash,
-            mask_email(normalized_recipient),
-            exc,
-        )
-        return jsonify({"fel": "Ett internt fel har inträffat."}), 500
-
-    logger.info(
-        "Användare %s delade intyg %s med %s",
-        pnr_hash,
-        pdf_ids,
-        mask_email(normalized_recipient),
-    )
-    success_message = (
-        "Intyget har skickats via e-post."
-        if len(attachments) == 1
-        else "Intygen har skickats via e-post."
-    )
-    return jsonify({"meddelande": success_message}), 200
-
-
-@app.route("/view_pdf/<int:pdf_id>")
-def view_pdf(pdf_id: int):
-    # Redirect to a direct download of the specified PDF.
-    if not session.get("user_logged_in"):
-        logger.debug("Oautentiserat visningsförsök för %s", pdf_id)
-        return redirect("/login")
-    pnr_hash = session.get("personnummer")
-    if not pnr_hash:
-        return redirect("/login")
-    pdf = functions.get_pdf_metadata(pnr_hash, pdf_id)
-    if not pdf:
-        logger.warning("PDF %s hittades inte för användare %s", pdf_id, pnr_hash)
-        abort(404)
-    logger.info("Användare %s laddar ned %s via direktlänk", pnr_hash, pdf["filename"])
-    return redirect(url_for("download_pdf", pdf_id=pdf_id))
-
-
-@app.route("/admin", methods=["POST", "GET"])
-def admin():  # pragma: no cover
-    # Admin dashboard for uploading certificates and creating users.
-    if request.method == "POST":
-        if not session.get("admin_logged_in"):
-            logger.warning("Obehörigt admin-POST-anrop")
-            return redirect("/error", code=401)
-
-        try:
-            # --- Grab form data ---
-            email_input = request.form.get("email", "").strip()
-            username = request.form.get("username", "").strip()
-            personnummer = functions.normalize_personnummer(
-                request.form.get("personnummer", "").strip()
-            )
-            normalized_email = functions.normalize_email(email_input)
-            email = normalized_email
-            email_hash = functions.hash_value(email)
-            pnr_hash = functions.hash_value(personnummer)
-
-            raw_categories = request.form.getlist("categories")
-            pdf_files = request.files.getlist("pdf")
-            if not raw_categories:
-                logger.warning("Adminuppladdning saknar kategorier (inget val)")
-                return jsonify({"status": "error", "message": "Välj kategori för varje PDF."}), 400
-            if not pdf_files:
-                logger.warning("Adminuppladdning utan PDF")
-                return jsonify({"status": "error", "message": "PDF-fil saknas"}), 400
-
-            if len(raw_categories) != len(pdf_files):
-                logger.warning(
-                    "Adminuppladdning med mismatch i kategorier och filer (kategorier=%d, filer=%d)",
-                    len(raw_categories),
-                    len(pdf_files),
-                )
-                return jsonify({"status": "error", "message": "Välj kategori för varje PDF."}), 400
-
-            logger.debug(
-                "Adminuppladdning för %s med kategorier %s",
-                mask_hash(pnr_hash),
-                raw_categories,
-            )
-
-            normalized_categories = []
-            for idx, raw in enumerate(raw_categories):
-                selected = normalize_category_slugs([raw])
-                if len(selected) != 1:
-                    logger.warning(
-                        "Adminuppladdning med ogiltig kategori för fil %d (värde=%r)",
-                        idx,
-                        raw,
-                    )
-                    return jsonify(
-                        {"status": "error", "message": "Välj giltig kategori för varje PDF."}
-                    ), 400
-                normalized_categories.append(selected[0])
-
-            # --- Check if user exists ---
-            user_exists = functions.get_user_info(personnummer) or functions.check_user_exists(
-                email
-            )
-            pending_exists = functions.check_pending_user_hash(pnr_hash)
-
-            # --- Save PDFs ---
-            pdf_records = [
-                pdf.save_pdf_for_user(personnummer, file_storage, [category], logger=logger)
-                for file_storage, category in zip(pdf_files, normalized_categories)
-            ]
-
-            # --- Return early for existing or pending users ---
-            if user_exists:
-                logger.info(
-                    "PDF:er uppladdade för befintlig användare %s (%d filer)",
-                    mask_hash(pnr_hash),
-                    len(pdf_records),
-                )
-                return jsonify(
-                    {
-                        "status": "success",
-                        "message": "PDF:er uppladdade för befintligt standardkonto",
-                    }
-                )
-
-            if pending_exists:
-                logger.info(
-                    "PDF:er uppladdade för väntande användare %s (%d filer)",
-                    mask_hash(pnr_hash),
-                    len(pdf_records),
-                )
-                return jsonify(
-                    {
-                        "status": "success",
-                        "message": "Standardkontot väntar redan på aktivering. PDF:er uppladdade.",
-                    }
-                )
-
-            # --- Create new pending user ---
-            if functions.admin_create_user(email, username, personnummer):
-                link = url_for("create_user", pnr_hash=pnr_hash, _external=True)
-                try:
-                    email_service.send_creation_email(email, link)
-                except RuntimeError as e:
-                    logger.error(
-                        "Misslyckades med att skicka skapandemejl till %s",
-                        mask_hash(email_hash),
-                        exc_info=True,
-                    )
-                    return jsonify(
-                        {
-                            "status": "error",
-                            "message": "Det gick inte att skicka inloggningslänken via e-post.",
-                        }
-                    ), 500
-
-                logger.info("Admin skapade användare %s", mask_hash(pnr_hash))
-                return jsonify(
-                    {"status": "success", "message": "Standardkonto skapat", "link": link}
-                )
-
-            logger.error("Misslyckades med att skapa väntande användare för %s", mask_hash(pnr_hash))
-            return jsonify({"status": "error", "message": "Kunde inte skapa standardkonto"}), 500
-
-        except ValueError as ve:
-            logger.error("Värdefel under adminuppladdning: %s", ve)
-            return jsonify({"status": "error", "message": "Felaktiga användardata."}), 400
-        except RequestEntityTooLarge:
-            raise
-        except Exception as e:
-            logger.error("Serverfel under adminuppladdning", exc_info=e)
-            return jsonify({"status": "error", "message": "Serverfel"}), 500
-
-    # --- GET request ---
-    if not session.get("admin_logged_in"):
-        logger.warning("Obehörigt admin-GET-anrop")
-        return redirect("/login_admin")
-
-    admin_log_entries = []
-    try:
-        with functions.get_engine().connect() as conn:
-            rows = conn.execute(
-                functions.admin_audit_log_table.select()
-                .order_by(functions.admin_audit_log_table.c.created_at.desc())
-                .limit(10)
-            ).fetchall()
-        for row in rows:
-            created_at = row.created_at
-            if created_at and hasattr(created_at, "isoformat"):
-                created_at = created_at.isoformat(sep=" ", timespec="minutes")
-            admin_log_entries.append(
-                {
-                    "admin": row.admin,
-                    "action": row.action,
-                    "details": row.details,
-                    "created_at": created_at or "",
-                }
-            )
-    except Exception:
-        logger.error("Misslyckades att hämta adminlogg")
-
-    logger.debug("Renderar adminsida")
-    return render_template(
-        "admin.html",
-        admin_log_entries=admin_log_entries,
-    )
-
-
-@app.get("/admin/guide")
-def admin_guide():  # pragma: no cover
-    if not session.get("admin_logged_in"):
-        logger.warning("Obehörigt admin guide-GET-anrop")
-        return redirect("/login_admin")
-    guide_path = Path(current_app.root_path) / "admin.md"
-    try:
-        guide_content = guide_path.read_text(encoding="utf-8")
-    except FileNotFoundError:
-        logger.error("Admin guide-fil saknas")
-        guide_content = "Guiden kunde inte hittas."
-    rendered_guide = render_markdown_content(guide_content)
-    logger.debug("Renderar admin guide-sida")
-    return render_template("admin_guide.html", guide_content=rendered_guide)
-
-
-@app.get("/admin/konton")
-def admin_accounts():  # pragma: no cover
-    if not session.get("admin_logged_in"):
-        logger.warning("Obehörigt admin accounts-GET-anrop")
-        return redirect("/login_admin")
-    csrf_token = ensure_csrf_token()
-    logger.debug("Renderar admin accounts-sida")
-    return render_template(
-        "admin_accounts.html",
-        categories=COURSE_CATEGORIES,
-        category_groups=COURSE_CATEGORY_GROUPS,
-        csrf_token=csrf_token,
-    )
-
-
-@app.get("/admin/intyg")
-def admin_certificates():  # pragma: no cover
-    if not session.get("admin_logged_in"):
-        logger.warning("Obehörigt admin certificates-GET-anrop")
-        return redirect("/login_admin")
-    logger.debug("Renderar admin certificates-sida")
-    return render_template(
-        "admin_certificates.html",
-        categories=COURSE_CATEGORIES,
-    )
-
-
-@app.get("/admin/foretagskonto")
-def admin_company_accounts():  # pragma: no cover
-    if not session.get("admin_logged_in"):
-        logger.warning("Obehörigt admin company accounts-GET-anrop")
-        return redirect("/login_admin")
-    logger.debug("Renderar admin company accounts-sida")
-    csrf_token = ensure_csrf_token()
-    return render_template("admin_company_accounts.html", csrf_token=csrf_token)
-
-
-def _send_application_approval_emails(
-    application_id: int, result: dict[str, Any]
-) -> dict[str, str]:
-    payload: dict[str, str] = {"message": "Ansökan godkänd."}
-    email_warnings: list[str] = []
-
-    if result.get("user_activation_required") and result.get("user_personnummer_hash"):
-        link = url_for(
-            "create_user",
-            pnr_hash=result["user_personnummer_hash"],
-            _external=True,
-        )
-        try:
-            email_service.send_creation_email(result["email"], link)
-            payload["message"] = (
-                "Ansökan godkänd. Aktiveringsmejl skickat till standardkontot."
-            )
-            payload["access_link"] = link
-            payload["access_link_label"] = "Aktiveringslänk"
-            payload["access_link_type"] = "activation"
-            payload["creation_link"] = link
-        except Exception:
-            logger.exception(
-                "Misslyckades att skicka aktiveringslänk till sökande för ansökan %s",
-                application_id,
-            )
-            email_warnings.append("Aktiveringslänken till sökande kunde inte skickas.")
-
-    if result.get("account_type") == "foretagskonto":
-        supervisor_email = result.get("supervisor_email") or result.get("email")
-        if result.get("supervisor_activation_required") and supervisor_email:
-            try:
-                activation_token = functions.ensure_pending_supervisor_activation_token(
-                    supervisor_email
-                )
-                link = url_for(
-                    "supervisor_create",
-                    activation_token=activation_token,
-                    _external=True,
-                )
-                email_service.send_creation_email(supervisor_email, link)
-                payload["message"] = (
-                    "Ansökan godkänd. Aktiveringsmejl skickat till företagskontot."
-                )
-                payload["access_link"] = link
-                payload["access_link_label"] = "Aktiveringslänk"
-                payload["access_link_type"] = "activation"
-                payload["creation_link"] = link
-            except Exception:
-                logger.exception(
-                    "Misslyckades att skicka aktiveringslänk till supervisor (ansökan %s)",
-                    application_id,
-                )
-                email_warnings.append(
-                    "Aktiveringslänken till företagskontot kunde inte skickas."
-                )
-        elif supervisor_email:
-            try:
-                reset_token = functions.create_supervisor_password_reset_token(
-                    supervisor_email
-                )
-                link = url_for(
-                    "supervisor_password_reset",
-                    token=reset_token,
-                    _external=True,
-                )
-                email_service.send_password_reset_email(supervisor_email, link)
-                payload["message"] = (
-                    "Ansökan godkänd. Företagskontot finns redan och ett åtkomstmejl "
-                    "har skickats."
-                )
-                payload["access_link"] = link
-                payload["access_link_label"] = "Återställningslänk"
-                payload["access_link_type"] = "password_reset"
-            except Exception:
-                logger.exception(
-                    "Misslyckades att skicka åtkomstmejl till aktivt företagskonto "
-                    "för ansökan %s",
-                    application_id,
-                )
-                email_warnings.append(
-                    "Åtkomstmejlet till företagskontot kunde inte skickas."
-                )
-
-    if email_warnings:
-        payload["email_warning"] = " ".join(email_warnings)
-
-    return payload
-
-
-@app.route("/admin/ansokningar", methods=["GET", "POST"])
-def admin_applications():  # pragma: no cover
-    if not session.get("admin_logged_in"):
-        return redirect("/login_admin")
-    if request.method == "POST":
-        if not validate_csrf_token(allow_if_absent=True):
-            return jsonify(
-                {
-                    "status": "error",
-                    "message": "Formuläret är inte längre giltigt. Ladda om sidan och försök igen.",
-                }
-            ), 400
-        application_id = request.form.get("application_id")
-        application_status = request.form.get("application_status")
-        if not application_id:
-            return jsonify({"status": "error", "message": "Ansökans ID saknas."}), 400
-        if application_status == "approved":
-            result = functions.approve_application_request(int(application_id), "admin")
-            email_result = _send_application_approval_emails(
-                int(application_id), result
-            )
-            logger.debug("Ansökan har godkänts, lyckat")
-            payload: dict[str, str] = {
-                "status": "success",
-                "message": email_result["message"],
-            }
-            for key in (
-                "email_warning",
-                "creation_link",
-                "access_link",
-                "access_link_label",
-                "access_link_type",
-            ):
-                value = email_result.get(key)
-                if value:
-                    payload[key] = value
-            return jsonify(payload)
-        elif application_status == "rejected":
-            functions.reject_application_request(int(application_id), "admin")
-            logger.debug("Ansökan har avslagits, lyckat")
-            return jsonify({"status": "success", "message": "Ansökan har avslagits."})
-        else:
-            logger.error("Ogiltig status för ansökan.")
-            return jsonify({"status": "error", "message": "Ogiltig status för ansökan."}), 400
-    elif request.method == "GET":
-        applications_requests = functions.list_application_requests()
-
-        if logger.isEnabledFor(logging.DEBUG):
-
-            def _mask_text(value: Any) -> str:
-                # Mask free-text values for logging to avoid PII leakage.
-                if value is None:
-                    return mask_sensitive_data(value)
-                text_value = str(value).strip()
-                if not text_value:
-                    return mask_sensitive_data(text_value)
-                return mask_hash(functions.hash_value(text_value))
-
-            for application in applications_requests:
-                logger.debug(
-                    "ID: %s, Typ: %s, Namn: %s, E-post: %s, OrgNr: %s, "
-                    "Företagsnamn: %s, Fakturaadress: %s, Fakturakontakt: %s, "
-                    "Fakturareferens: %s, Kommentar: %s, Status: %s, Granskad av: %s, "
-                    "Beslutsorsak: %s, Skapad: %s, Uppdaterad: %s, Granskad: %s",
-                    application.get("id"),
-                    application.get("account_type"),
-                    _mask_text(application.get("name")),
-                    mask_email(application.get("email", "")),
-                    _mask_text(application.get("orgnr_normalized")),
-                    _mask_text(application.get("company_name")),
-                    _mask_text(application.get("invoice_address")),
-                    _mask_text(application.get("invoice_contact")),
-                    _mask_text(application.get("invoice_reference")),
-                    _mask_text(application.get("comment")),
-                    application.get("status"),
-                    _mask_text(application.get("reviewed_by")),
-                    _mask_text(application.get("decision_reason")),
-                    application.get("created_at"),
-                    application.get("updated_at"),
-                    application.get("reviewed_at"),
-                )
-        csrf_token = sec.ensure_csrf_token()
-        return render_template(
-            "admin_applications.html", applications=applications_requests, csrf_token=csrf_token
-        )
-    else:
-        return render_template("error.html", code=405, message="Method Not Allowed"), 405
-
-
-@app.route("/gdpr", methods=["GET"])
-def gdpr_info():
-    return render_template("gdpr.html")
-
-
-@app.route("/admin/fakturering", methods=["GET"])
-def admin_invoicing():  # pragma: no cover
-    if not session.get("admin_logged_in"):
-        return redirect("/login_admin")
-    companies = functions.list_companies_for_invoicing()
-    return render_template("admin_invoicing.html", companies=companies)
-
-
-def _serialize_application_row(row: dict) -> dict:  # pragma: no cover
-    def _safe_isoformat(value: Any) -> str | None:
-        if value is None:
-            return None
-        if hasattr(value, "isoformat"):
-            return value.isoformat()
-        return None
-
-    return {
-        "id": row.get("id"),
-        "account_type": row.get("account_type"),
-        "name": row.get("name"),
-        "email": row.get("email"),
-        "orgnr_normalized": row.get("orgnr_normalized"),
-        "company_name": row.get("company_name"),
-        "invoice_address": row.get("invoice_address"),
-        "invoice_contact": row.get("invoice_contact"),
-        "invoice_reference": row.get("invoice_reference"),
-        "comment": row.get("comment"),
-        "status": row.get("status"),
-        "reviewed_by": row.get("reviewed_by"),
-        "decision_reason": row.get("decision_reason"),
-        "created_at": _safe_isoformat(row.get("created_at")),
-        "updated_at": _safe_isoformat(row.get("updated_at")),
-        "reviewed_at": _safe_isoformat(row.get("reviewed_at")),
-    }
-
-
-@app.get("/admin/api/ansokningar")
-def admin_list_applications():  # pragma: no cover
-    _require_admin()
-    status = request.args.get("status")
-    try:
-        rows = functions.list_application_requests(status)
-    except ValueError as exc:
-        logger.warning(
-            "Invalid status in list_application_requests: %s",
-            exc,
-        )
-        logger.exception("Misslyckades med att lista ansökningar")
-        return jsonify({"status": "error", "message": "Felaktig begäran."}), 400
-
-    serialized = [_serialize_application_row(row) for row in rows]
-    return jsonify({"status": "success", "data": serialized})
-
-
-@app.get("/admin/api/ansokningar/<int:application_id>")
-def admin_get_application(application_id: int):  # pragma: no cover
-    _require_admin()
-    row = functions.get_application_request(application_id)
-    if not row:
-        return jsonify({"status": "error", "message": "Ansökan hittades inte."}), 404
-    return jsonify({"status": "success", "data": _serialize_application_row(row)})
-
-
-@app.post("/admin/api/ansokningar/<int:application_id>/godkann")
-def admin_approve_application(application_id: int):  # pragma: no cover
-    """
-    Godkänn ansökan, skapa/aktivera konto(n), skicka mejl till sökande (normal user)
-    och – vid företagskonto – även till handledare/supervisor. Logga admin-åtgärden.
-    """
-    admin_name = _require_admin()
-    if not validate_csrf_token():
-        return _api_error_response(
-            "Ogiltig CSRF-token.",
-            400,
-            severity="warning",
-            extra={"route": "admin_approve_application", "admin": admin_name},
-        )
-
-    try:
-        # Förväntat result-format (exempel):
-        # {
-        #   'email': 'applicant@example.com',
-        #   'account_type': 'standard'|'foretagskonto',
-        #   'company_name': 'Exempel AB'|None,
-        #   'user_activation_required': True|False,
-        #   'user_personnummer_hash': 'abc123'|None,
-        #   'supervisor_activation_required': True|False,
-        #   'supervisor_email_hash': 'def456'|None
-        # }
-        result = functions.approve_application_request(application_id, admin_name)
-    except ValueError as exc:
-        return _api_error_response(
-            _safe_user_error(
-                "",
-                ALLOWED_ADMIN_APPROVAL_ERRORS,
-                (
-                    "Ansökan kunde inte godkännas. Kontrollera att den fortfarande "
-                    "är väntande och att uppgifterna är kompletta."
-                ),
-            ),
-            400,
-            severity="warning",
-            error=exc,
-            extra={
-                "route": "admin_approve_application",
-                "application_id": application_id,
-                "admin": admin_name,
-            },
-        )
-    except Exception as exc:
-        return _api_error_response(
-            "Kunde inte godkänna ansökan.",
-            500,
-            severity="error",
-            error=exc,
-            extra={
-                "route": "admin_approve_application",
-                "application_id": application_id,
-                "admin": admin_name,
-            },
-        )
-
-    email_result = _send_application_approval_emails(application_id, result)
-
-    masked_email = mask_hash(functions.hash_value(result["email"]))
-    functions.log_admin_action(
-        admin_name,
-        "godkände ansökan",
-        f"application_id={application_id}, email={masked_email}",
-    )
-
-    payload: dict[str, Any] = {
-        "status": "success",
-        "data": result,
-        "message": email_result["message"],
-    }
-    for key in (
-        "email_warning",
-        "creation_link",
-        "access_link",
-        "access_link_label",
-        "access_link_type",
-    ):
-        value = email_result.get(key)
-        if value:
-            payload[key] = value
-
-    return jsonify(payload)
-
-
-@app.post("/admin/api/ansokningar/<int:application_id>/avslag")
-def admin_reject_application(application_id: int):  # pragma: no cover
-    admin_name = _require_admin()
-    if not validate_csrf_token():
-        return _api_error_response(
-            "Ogiltig CSRF-token.",
-            400,
-            severity="warning",
-            extra={"route": "admin_reject_application", "admin": admin_name},
-        )
-
-    payload = request.get_json(silent=True) or {}
-    decision_reason = (payload.get("reason") or "").strip()
-    if not decision_reason:
-        decision_reason = "Ingen motivering angiven."
-
-    try:
-        result = functions.reject_application_request(application_id, admin_name, decision_reason)
-    except ValueError as exc:
-        return _api_error_response(
-            (
-                "Ansökan kunde inte avslås. Kontrollera att den fortfarande "
-                "är väntande och försök igen."
-            ),
-            400,
-            severity="warning",
-            error=exc,
-            extra={
-                "route": "admin_reject_application",
-                "application_id": application_id,
-                "admin": admin_name,
-            },
-        )
-    except Exception as exc:
-        return _api_error_response(
-            "Kunde inte avslå ansökan.",
-            500,
-            severity="error",
-            error=exc,
-            extra={
-                "route": "admin_reject_application",
-                "application_id": application_id,
-                "admin": admin_name,
-            },
-        )
-
-    email_error = None
-    try:
-        email_service.send_application_rejection_email(
-            result["email"], result["company_name"], decision_reason
-        )
-    except RuntimeError as exc:
-        logger.error("Misslyckades att skicka avslag för ansökan %s", application_id)
-        email_error = str(exc)
-
-    masked_email = mask_hash(functions.hash_value(result["email"]))
-    functions.log_admin_action(
-        admin_name,
-        "avslog ansökan",
-        f"application_id={application_id}, email={masked_email}",
-    )
-
-    response_payload = {"status": "success", "data": result}
-    if email_error:
-        response_payload["email_warning"] = "Ansökan avslogs men e-post kunde inte skickas."
-    return jsonify(response_payload)
-
-
-@app.post("/admin/api/oversikt")
-def admin_user_overview():  # pragma: no cover
-    admin_name = _require_admin()
-    payload = request.get_json(silent=True) or {}
-    personnummer = (payload.get("personnummer") or "").strip()
-    if not personnummer:
-        logger.debug("Admin overview without personnummer", extra={"admin": admin_name})
-        return jsonify({"status": "error", "message": "Ange personnummer."}), 400
-    try:
-        normalized_personnummer = functions.normalize_personnummer(personnummer)
-    except ValueError:
-        personnummer_masked = mask_hash(functions.hash_value(personnummer))
-        logger.debug(
-            "Admin overview with invalid personnummer hash: %s",
-            personnummer_masked,
-            extra={"admin": admin_name},
-        )
-        return jsonify({"status": "error", "message": "Ogiltigt personnummer."}), 400
-
-    pnr_hash = functions.hash_value(normalized_personnummer)
-    pdfs = functions.get_user_pdfs(pnr_hash)
-    overview = []
-    for pdf in pdfs:
-        uploaded_at = pdf.get("uploaded_at")
-        expires_on = pdf.get("expires_on")
-        overview.append(
-            {
-                "id": pdf["id"],
-                "filename": pdf["filename"],
-                "categories": pdf.get("categories") or [],
-                "category_labels": labels_for_slugs(pdf.get("categories") or []),
-                "expires_on": expires_on.isoformat()
-                if (expires_on and hasattr(expires_on, "isoformat"))
-                else None,
-                "uploaded_at": uploaded_at.isoformat()
-                if (uploaded_at and hasattr(uploaded_at, "isoformat"))
-                else None,
-            }
-        )
-
-    user_row = functions.get_user_info(normalized_personnummer)
-    pending = functions.check_pending_user(normalized_personnummer)
-    response = {
-        "status": "success",
-        "data": {
-            "personnummer_hash": pnr_hash,
-            "username": user_row.username if user_row else None,
-            "email_hash": user_row.email if user_row else None,
-            "pending": pending,
-            "pdfs": overview,
-            "categories": [{"slug": slug, "label": label} for slug, label in COURSE_CATEGORIES],
-        },
-    }
-    functions.log_admin_action(
-        admin_name,
-        "visade användaröversikt",
-        f"personnummer_hash={pnr_hash}",
-    )
-    logger.debug(
-        "Admin overview for %s with %d pdfs",
-        mask_hash(pnr_hash),
-        len(pdfs),
-        extra={"admin": admin_name},
-    )
-    return jsonify(response)
-
-
-@app.post("/admin/api/klientlogg")
-def admin_client_log():
-    admin_name = _require_admin()
-    payload = request.get_json(silent=True)
-    if not isinstance(payload, dict):
-        return jsonify({"status": "error", "message": "Ogiltigt loggformat."}), 400
-    message = _truncate_log_value(payload.get("message"), CLIENT_LOG_TRUNCATION_LIMITS["message"])
-    context = _truncate_log_value(payload.get("context"), CLIENT_LOG_TRUNCATION_LIMITS["context"])
-    url = _truncate_log_value(payload.get("url"), CLIENT_LOG_TRUNCATION_LIMITS["url"])
-    details = payload.get("details")
-    if isinstance(details, (dict, list)):
-        masked_details = mask_sensitive_data(details)
-        masked_details = _truncate_log_value(
-            json.dumps(masked_details, ensure_ascii=False),
-            CLIENT_LOG_TRUNCATION_LIMITS["details"],
-        )
-    else:
-        masked_details = _truncate_log_value(details, CLIENT_LOG_TRUNCATION_LIMITS["details"])
-    status = payload.get("status")
-    logger.warning(
-        "Klientlogg från admin %s: %s | context=%s | url=%s | status=%s | details=%s",
-        admin_name,
-        message or "Okänt fel",
-        context or "okänd",
-        url or "okänd",
-        status,
-        masked_details,
-    )
-    return jsonify({"status": "success"})
-
-
-@app.post("/admin/api/radera-pdf")
-def admin_delete_pdf():  # pragma: no cover
-    admin_name = _require_admin()
-    payload = request.get_json(silent=True) or {}
-    personnummer = (payload.get("personnummer") or "").strip()
-    pdf_id = payload.get("pdf_id")
-    if not personnummer or pdf_id is None:
-        logger.debug(
-            "Admin delete_pdf without personnummer or pdf_id", extra={"admin": admin_name}
-        )
-        return jsonify({"status": "error", "message": "Ange personnummer och PDF-id."}), 400
-    try:
-        normalized_personnummer = functions.normalize_personnummer(personnummer)
-    except ValueError:
-        personnummer_masked = mask_hash(functions.hash_value(personnummer))
-        logger.debug(
-            "Admin delete_pdf with invalid personnummer hash: %s",
-            personnummer_masked,
-            extra={"admin": admin_name},
-        )
-        return jsonify({"status": "error", "message": "Ogiltigt personnummer."}), 400
-
-    try:
-        pdf_id_int = int(pdf_id)
-    except (ValueError, TypeError):
-        logger.debug(
-            "Admin delete_pdf with invalid pdf_id: %s", pdf_id, extra={"admin": admin_name}
-        )
-        return jsonify({"status": "error", "message": "Ogiltigt PDF-id."}), 400
-
-    if not functions.delete_user_pdf(normalized_personnummer, pdf_id_int):
-        return (
-            jsonify({"status": "error", "message": "PDF kunde inte hittas."}),
-            404,
-        )
-
-    pnr_hash = functions.hash_value(normalized_personnummer)
-    functions.log_admin_action(
-        admin_name,
-        "raderade PDF",
-        f"personnummer_hash={pnr_hash}, pdf_id={pdf_id_int}",
-    )
-    logger.info(
-        "Admin deleted pdf %s for %s", pdf_id_int, mask_hash(pnr_hash), extra={"admin": admin_name}
-    )
-    return jsonify({"status": "success", "message": "PDF borttagen."})
-
-
-@app.post("/admin/api/radera-konto")
-def admin_delete_account():  # pragma: no cover
-    admin_name = _require_admin()
-    if not validate_csrf_token():
-        return jsonify({"status": "error", "message": "Ogiltig CSRF-token."}), 400
-    payload = request.get_json(silent=True) or {}
-    personnummer = (payload.get("personnummer") or "").strip()
-    personnummer_hash = (payload.get("personnummer_hash") or "").strip()
-    notify_email = (payload.get("email") or "").strip()
-    if not personnummer:
-        if not personnummer_hash:
-            logger.debug(
-                "Admin delete_account without personnummer",
-                extra={"admin": admin_name},
-            )
-            return jsonify({"status": "error", "message": "Välj ett konto att radera."}), 400
-    personnummer_masked = (
-        mask_hash(functions.hash_value(personnummer)) if personnummer else "saknas"
-    )
-    try:
-        if personnummer:
-            normalized_personnummer = functions.normalize_personnummer(personnummer)
-            personnummer_hash = functions.hash_value(normalized_personnummer)
-        elif not functions._is_valid_hash(personnummer_hash):
-            raise ValueError("Ogiltig hash")
-    except ValueError:
-        logger.debug(
-            "Admin delete_account with invalid personnummer: %s",
-            personnummer_masked,
-            extra={"admin": admin_name},
-        )
-        return jsonify({"status": "error", "message": "Ogiltiga uppgifter."}), 400
-
-    normalized_email = None
-    if notify_email:
-        try:
-            normalized_email = email_service.normalize_valid_email(notify_email)
-        except ValueError:
-            return jsonify({"status": "error", "message": "Ogiltig e-postadress."}), 400
-
-    try:
-        deleted, summary, username = functions.admin_delete_user_account_by_hash(personnummer_hash)
-    except Exception:
-        logger.error(
-            "Misslyckades att radera konto för %s",
-            personnummer_masked,
-        )
-        return jsonify({"status": "error", "message": "Kunde inte radera kontot."}), 500
-
-    if not deleted:
-        return jsonify({"status": "error", "message": "Kontot hittades inte."}), 404
-
-    email_warning = None
-    if normalized_email:
-        try:
-            email_service.send_account_deletion_email(normalized_email, username)
-        except Exception:
-            logger.error("Misslyckades att skicka raderingsmejl")
-            email_warning = "Kontot raderades, men mejlet kunde inte skickas."
-
-    pnr_hash = personnummer_hash
-    summary_details = ", ".join(f"{key}={value}" for key, value in summary.items())
-    email_hash = (
-        functions.hash_value(functions.normalize_email(notify_email))
-        if normalized_email
-        else "saknas"
-    )
-    functions.log_admin_action(
-        admin_name,
-        "raderade konto",
-        f"personnummer_hash={pnr_hash}, email_hash={email_hash}, {summary_details}",
-    )
-    logger.info(
-        "Admin deleted account for %s",
-        mask_hash(pnr_hash),
-        extra={"admin": admin_name},
-    )
-    response_payload = {
-        "status": "success",
-        "message": (
-            "Kontot har raderats och mejl har skickats."
-            if normalized_email
-            else "Kontot har raderats."
-        ),
-        "data": summary,
-    }
-    if email_warning:
-        response_payload["email_warning"] = email_warning
-        response_payload["message"] = "Kontot har raderats."
-    return jsonify(response_payload)
-
-
-@app.get("/admin/api/konton/lista")
-def admin_list_accounts():  # pragma: no cover
-    _require_admin()
-    accounts = functions.list_admin_accounts()
-    return jsonify({"status": "success", "data": accounts})
-
-
-@app.get("/admin/api/epost-hashar/lista")
-def admin_list_legacy_email_hashes():  # pragma: no cover
-    _require_admin()
-    references = functions.list_legacy_email_references()
-    return jsonify({"status": "success", "data": references})
-
-
-@app.post("/admin/api/epost-hashar/komplettera")
-def admin_complete_legacy_email_hash():  # pragma: no cover
-    admin_name = _require_admin()
-    payload = request.get_json(silent=True)
-    if payload is None:
-        payload = {}
-    elif not isinstance(payload, dict):
-        return jsonify({"status": "error", "message": "Ogiltig begäran."}), 400
-    if not validate_csrf_token():
-        return jsonify({"status": "error", "message": "Ogiltig CSRF-token."}), 400
-    reference_type = (payload.get("reference_type") or "").strip()
-    email_hash = (payload.get("email_hash") or "").strip()
-    email = (payload.get("email") or "").strip()
-    personnummer_hash = (payload.get("personnummer_hash") or "").strip() or None
-    if not reference_type or not email_hash or not email:
-        return jsonify({"status": "error", "message": "Välj referens och ange e-post."}), 400
-
-    try:
-        success, summary, error = functions.complete_legacy_email_reference(
-            reference_type,
-            email_hash,
-            email,
-            personnummer_hash,
-        )
-    except ValueError:
-        return jsonify({"status": "error", "message": "Ogiltig e-postadress."}), 400
-
-    if not success:
-        messages = {
-            "invalid_hash": "Ogiltig e-postreferens.",
-            "invalid_personnummer": "Ogiltig personnummerreferens.",
-            "invalid_type": "Ogiltig referenstyp.",
-            "hash_mismatch": "E-postadressen matchar inte den valda hashen.",
-            "missing": "Referensen hittades inte.",
-            "email_in_use": "E-postadressen används redan av ett annat konto.",
-        }
-        status_code = 404 if error == "missing" else 409 if error == "email_in_use" else 400
-        return (
-            jsonify(
-                {
-                    "status": "error",
-                    "message": messages.get(error or "", "Kunde inte komplettera e-post."),
-                }
-            ),
-            status_code,
-        )
-
-    summary = summary or {}
-    summary_details = ", ".join(f"{key}={value}" for key, value in summary.items())
-    functions.log_admin_action(
-        admin_name,
-        "kompletterade e-posthash",
-        (
-            f"typ={reference_type}, email_ref={mask_hash(email_hash)}, "
-            f"epost={mask_email(functions.normalize_email(email))}, {summary_details}"
-        ),
-    )
-    return jsonify(
-        {
-            "status": "success",
-            "message": "E-postadressen har kompletterats.",
-            "data": summary,
-        }
-    )
-
-
-@app.post("/admin/api/konton/uppdatera")
-def admin_update_account():  # pragma: no cover
-    admin_name = _require_admin()
-    if not validate_csrf_token():
-        return jsonify({"status": "error", "message": "Ogiltig CSRF-token."}), 400
-    payload = request.get_json(silent=True) or {}
-    personnummer = (payload.get("personnummer") or "").strip()
-    email = (payload.get("email") or "").strip()
-    username = (payload.get("username") or "").strip()
-    if not personnummer or not email or not username:
-        return jsonify({"status": "error", "message": "Fyll i personnummer, namn och e-post."}), 400
-
-    try:
-        normalized_personnummer = functions.normalize_personnummer(personnummer)
-        normalized_email = functions.normalize_email(email)
-    except ValueError:
-        return jsonify({"status": "error", "message": "Ogiltiga uppgifter."}), 400
-
-    success, summary, error = functions.admin_update_user_account(
-        normalized_personnummer, normalized_email, username
-    )
-    if not success:
-        if error == "missing_account":
-            return jsonify({"status": "error", "message": "Kontot hittades inte."}), 404
-        if error == "email_in_use":
-            return jsonify({"status": "error", "message": "E-postadressen används redan."}), 409
-        return jsonify({"status": "error", "message": "Kunde inte uppdatera kontot."}), 400
-    if summary is None:
-        return jsonify({"status": "error", "message": "Kunde inte uppdatera kontot."}), 400
-
-    pnr_hash = functions.hash_value(normalized_personnummer)
-    email_hash = functions.hash_value(normalized_email)
-    summary_details = ", ".join(f"{key}={value}" for key, value in summary.items())
-    functions.log_admin_action(
-        admin_name,
-        "uppdaterade konto",
-        f"personnummer_hash={pnr_hash}, email_hash={email_hash}, {summary_details}",
-    )
-    logger.info("Admin updated account for %s", mask_hash(pnr_hash), extra={"admin": admin_name})
-    return jsonify(
-        {
-            "status": "success",
-            "message": "Kontot har uppdaterats.",
-            "data": summary,
-        }
-    )
-
-
-@app.post("/admin/api/uppdatera-pdf")
-def admin_update_pdf():  # pragma: no cover
-    admin_name = _require_admin()
-    payload = request.get_json(silent=True) or {}
-    personnummer = (payload.get("personnummer") or "").strip()
-    pdf_id = payload.get("pdf_id")
-    categories = payload.get("categories")
-    if not isinstance(categories, list):
-        logger.debug(
-            "Admin update_pdf with invalid categories: %r", categories, extra={"admin": admin_name}
-        )
-        return jsonify({"status": "error", "message": "Kategorier måste vara en lista."}), 400
-    if not personnummer or pdf_id is None:
-        logger.debug(
-            "Admin update_pdf without personnummer or pdf_id", extra={"admin": admin_name}
-        )
-        return jsonify({"status": "error", "message": "Ange personnummer och PDF-id."}), 400
-    try:
-        normalized_personnummer = functions.normalize_personnummer(personnummer)
-    except ValueError:
-        personnummer_masked = mask_hash(functions.hash_value(personnummer))
-        logger.debug(
-            "Admin update_pdf with invalid personnummer hash: %s",
-            personnummer_masked,
-            extra={"admin": admin_name},
-        )
-        return jsonify({"status": "error", "message": "Ogiltigt personnummer."}), 400
-    try:
-        pdf_id_int = int(pdf_id)
-    except (ValueError, TypeError):
-        logger.debug(
-            "Admin update_pdf with invalid pdf_id: %s", pdf_id, extra={"admin": admin_name}
-        )
-        return jsonify({"status": "error", "message": "Ogiltigt PDF-id."}), 400
-
-    try:
-        normalized_categories = normalize_category_slugs(categories)
-    except ValueError:
-        logger.debug(
-            "Admin update_pdf with invalid categories: %r", categories, extra={"admin": admin_name}
-        )
-        return jsonify({"status": "error", "message": "Ogiltig kategori vald."}), 400
-
-    if not functions.update_pdf_categories(
-        normalized_personnummer, pdf_id_int, normalized_categories
-    ):
-        return (
-            jsonify({"status": "error", "message": "PDF kunde inte uppdateras."}),
-            404,
-        )
-
-    pnr_hash = functions.hash_value(normalized_personnummer)
-    functions.log_admin_action(
-        admin_name,
-        "uppdaterade PDF-kategorier",
-        f"personnummer_hash={pnr_hash}, pdf_id={pdf_id_int}, kategorier={';'.join(normalized_categories)}",
-    )
-    logger.info(
-        "Admin updated categories for pdf %s for %s",
-        pdf_id_int,
-        mask_hash(pnr_hash),
-        extra={"admin": admin_name},
-    )
-    return jsonify({"status": "success", "message": "Kategorier uppdaterade."})
-
-
-@app.post("/admin/api/konton/losenord-status")
-def admin_password_status():  # pragma: no cover
-    admin_name = _require_admin()
-    if not validate_csrf_token():
-        return jsonify({"status": "error", "message": "Ogiltig CSRF-token."}), 400
-    payload = request.get_json(silent=True) or {}
-    personnummer = (payload.get("personnummer") or "").strip()
-    email = (payload.get("email") or "").strip()
-
-    if not personnummer:
-        logger.debug(
-            "Admin password_status without personnummer",
-            extra={"admin": admin_name},
-        )
-        return jsonify({"status": "error", "message": "Ange personnummer."}), 400
-
-    try:
-        result = functions.get_admin_password_status(personnummer, email)
-    except ValueError:
-        return jsonify({"status": "error", "message": "Ogiltiga uppgifter."}), 400
-
-    if not result:
-        return (
-            jsonify(
-                {
-                    "status": "error",
-                    "message": "Uppgifterna matchar inget standardkonto eller väntande konto.",
-                }
-            ),
-            404,
-        )
-
-    if result["password_created"]:
-        message = "Lösenord är redan skapat för kontot."
-    else:
-        message = "Lösenord är inte skapat ännu. Skicka en skapa-konto-länk till användaren."
-
-    return jsonify(
-        {
-            "status": "success",
-            "message": message,
-            "data": result,
-        }
-    )
-
-
-@app.post("/admin/api/konton/skapa-losenordslank")
-def admin_send_create_password_link():  # pragma: no cover
-    admin_name = _require_admin()
-    if not validate_csrf_token():
-        return jsonify({"status": "error", "message": "Ogiltig CSRF-token."}), 400
-    payload = request.get_json(silent=True) or {}
-    personnummer = (payload.get("personnummer") or "").strip()
-    email = (payload.get("email") or "").strip()
-
-    if not personnummer:
-        logger.debug(
-            "Admin send_create_password_link without personnummer",
-            extra={"admin": admin_name},
-        )
-        return jsonify({"status": "error", "message": "Ange personnummer."}), 400
-
-    try:
-        personnummer_hash = functions.get_pending_user_personnummer_hash(personnummer)
-    except ValueError:
-        return jsonify({"status": "error", "message": "Ogiltiga uppgifter."}), 400
-
-    if not personnummer_hash:
-        return (
-            jsonify(
-                {
-                    "status": "error",
-                    "message": "Kontot är redan aktiverat eller uppgifterna matchar inget väntande konto.",
-                }
-            ),
-            404,
-        )
-
-    link = url_for("create_user", pnr_hash=personnummer_hash, _external=True)
-    if email:
-        try:
-            normalized_email = functions.normalize_email(email)
-        except ValueError:
-            logger.warning(
-                "Admin angav ogiltig e-postadress för skapa-konto-länk till %s",
-                mask_hash(personnummer_hash),
-                extra={"admin": admin_name},
-            )
-            return jsonify({"status": "error", "message": "Ogiltig e-postadress."}), 400
-        try:
-            email_service.send_creation_email(normalized_email, link)
-        except RuntimeError:
-            logger.error("Misslyckades att skicka skapa-konto-länk")
-            return jsonify({"status": "error", "message": "Kunde inte skicka skapa-konto-länk."}), 500
-
-        email_hash = functions.hash_value(normalized_email)
-        functions.log_admin_action(
-            admin_name,
-            "skickade skapa-konto-länk",
-            f"personnummer_hash={personnummer_hash}, email_hash={email_hash}",
-        )
-        logger.info(
-            "Admin sent create-password link for %s to %s",
-            mask_hash(personnummer_hash),
-            mask_hash(email_hash),
-            extra={"admin": admin_name},
-        )
-        message = "Skapa-konto-länk skickad."
-    else:
-        functions.log_admin_action(
-            admin_name,
-            "hämtade skapa-konto-länk",
-            f"personnummer_hash={personnummer_hash}",
-        )
-        logger.info(
-            "Admin fetched create-password link for %s without email send",
-            mask_hash(personnummer_hash),
-            extra={"admin": admin_name},
-        )
-        message = "Skapa-konto-länk hämtad. Ingen e-post angavs för utskick."
-
-    return jsonify(
-        {
-            "status": "success",
-            "message": message,
-            "link": link,
-        }
-    )
-
-
-@app.post("/admin/api/skicka-aterstallning")
-def admin_send_password_reset():  # pragma: no cover
-    admin_name = _require_admin()
-    payload = request.get_json(silent=True) or {}
-    account_type = (payload.get("account_type") or "standard").strip().lower()
-    personnummer = (payload.get("personnummer") or "").strip()
-    email = (payload.get("email") or "").strip()
-    if account_type not in {"standard", "foretagskonto"}:
-        logger.debug(
-            "Admin send_password_reset with invalid account_type: %s",
-            account_type,
-            extra={"admin": admin_name},
-        )
-
-    if account_type == "standard" and (not personnummer or not email):
-        logger.debug(
-            "Admin send_password_reset without personnummer or email", extra={"admin": admin_name}
-        )
-        return _api_error_response(
-            "Ange både personnummer och e-post.",
-            400,
-            severity="warning",
-            extra={"route": "admin_send_password_reset", "admin": admin_name},
-        )
-    if account_type == "foretagskonto" and not email:
-        logger.debug(
-            "Admin send_password_reset without email for foretagskonto", extra={"admin": admin_name}
-        )
-        return _api_error_response(
-            "Ange e-postadressen för företagskontot.",
-            400,
-            severity="warning",
-            extra={"route": "admin_send_password_reset", "admin": admin_name},
-        )
-
-    if account_type == "foretagskonto":
-        try:
-            token = functions.create_supervisor_password_reset_token(email)
-        except ValueError as exc:
-            return _api_error_response(
-                "Företagskontot hittades inte.",
-                404,
-                severity="warning",
-                error=exc,
-                extra={"route": "admin_send_password_reset", "admin": admin_name},
-            )
-        except Exception as exc:
-            return _api_error_response(
-                "Kunde inte skapa återställning.",
-                500,
-                severity="error",
-                error=exc,
-                extra={"route": "admin_send_password_reset", "admin": admin_name},
-            )
-
-        link = url_for("supervisor_password_reset", token=token, _external=True)
-        try:
-            email_service.send_password_reset_email(email, link)
-        except RuntimeError as exc:
-            return _api_error_response(
-                "Kunde inte skicka återställningsmejl.",
-                500,
-                severity="error",
-                error=exc,
-                extra={"route": "admin_send_password_reset", "admin": admin_name},
-            )
-
-        email_hash = functions.hash_value(functions.normalize_email(email))
-        functions.log_admin_action(
-            admin_name,
-            "skickade företagskonto-återställning",
-            f"email_hash={email_hash}",
-        )
-        logger.info(
-            "Admin sent supervisor password reset to %s",
-            mask_hash(email_hash),
-            extra={"admin": admin_name},
-        )
-        return jsonify(
-            {
-                "status": "success",
-                "message": "Återställningsmejl skickat till företagskontot.",
-                "link": link,
-            }
-        )
-
-    try:
-        normalized_personnummer = functions.normalize_personnummer(personnummer)
-    except ValueError as exc:
-        personnummer_masked = mask_hash(functions.hash_value(personnummer))
-        logger.debug(
-            "Admin send_password_reset with invalid personnummer hash: %s",
-            personnummer_masked,
-            extra={"admin": admin_name},
-        )
-        return _api_error_response(
-            "Ogiltigt personnummer.",
-            400,
-            severity="warning",
-            error=exc,
-            extra={"route": "admin_send_password_reset", "admin": admin_name},
-        )
-
-    try:
-        token = functions.create_password_reset_token(normalized_personnummer, email)
-    except ValueError as exc:
-        # Do not expose raw exception messages to the client; map to safe, user-facing text instead.
-        exc_message = str(exc)
-        if exc_message == "Kontot är inte aktiverat ännu.":
-            message = "Kontot är inte aktiverat ännu."
-            status_code = 409
-        else:
-            message = "Uppgifterna matchar inget aktivt standardkonto."
-            status_code = 404
-        return _api_error_response(
-            message,
-            status_code,
-            severity="warning",
-            error=exc,
-            extra={"route": "admin_send_password_reset", "admin": admin_name},
-        )
-    except Exception as exc:
-        return _api_error_response(
-            "Kunde inte skapa återställning.",
-            500,
-            severity="error",
-            error=exc,
-            extra={"route": "admin_send_password_reset", "admin": admin_name},
-        )
-
-    link = url_for("password_reset", token=token, _external=True)
-    try:
-        email_service.send_password_reset_email(email, link)
-    except RuntimeError as exc:
-        return _api_error_response(
-            "Kunde inte skicka återställningsmejl.",
-            500,
-            severity="error",
-            error=exc,
-            extra={"route": "admin_send_password_reset", "admin": admin_name},
-        )
-
-    pnr_hash = functions.hash_value(normalized_personnummer)
-    email_hash = functions.hash_value(functions.normalize_email(email))
-    functions.log_admin_action(
-        admin_name,
-        "skickade lösenordsåterställning",
-        f"personnummer_hash={pnr_hash}, email_hash={email_hash}",
-    )
-    logger.info(
-        "Admin sent password reset for %s to %s",
-        mask_hash(pnr_hash),
-        mask_hash(email_hash),
-        extra={"admin": admin_name},
-    )
-    return jsonify({"status": "success", "message": "Återställningsmejl skickat.", "link": link})
-
-
-@app.post("/admin/api/foretagskonto/skapa")
-def admin_create_supervisor_route():  # pragma: no cover
-    admin_name = _require_admin()
-    payload = request.get_json(silent=True) or {}
-    email = (payload.get("email") or "").strip()
-    name = (payload.get("name") or "").strip()
-    if not email or not name:
-        logger.debug("Admin create_supervisor without email or name", extra={"admin": admin_name})
-        return jsonify({"status": "error", "message": "Ange namn och e-post."}), 400
-
-    try:
-        normalized_email = functions.normalize_email(email)
-    except ValueError:
-        logger.debug(
-            "Admin create_supervisor with invalid email: %s",
-            mask_email(email),
-            extra={"admin": admin_name},
-        )
-        return jsonify({"status": "error", "message": "Ogiltig e-postadress."}), 400
-
-    if not functions.admin_create_supervisor(normalized_email, name):
-        return (
-            jsonify(
-                {
-                    "status": "error",
-                    "message": "Företagskontot finns redan eller väntar på aktivering.",
-                }
-            ),
-            409,
-        )
-
-    try:
-        activation_token = functions.ensure_pending_supervisor_activation_token(normalized_email)
-    except ValueError:
-        logger.error(
-            "Misslyckades med att skapa aktiveringslänk för handledare %s",
-            mask_email_reference(normalized_email),
-        )
-        return (
-            jsonify(
-                {
-                    "status": "error",
-                    "message": "Det gick inte att skapa inloggningslänken.",
-                }
-            ),
-            500,
-        )
-
-    link = url_for("supervisor_create", activation_token=activation_token, _external=True)
-
-    try:
-        email_service.send_creation_email(normalized_email, link)
-    except RuntimeError:
-        logger.error(
-            "Misslyckades med att skicka skapandemejl för handledare till %s",
-            mask_email_reference(normalized_email),
-        )
-        return (
-            jsonify(
-                {
-                    "status": "error",
-                    "message": "Det gick inte att skicka inloggningslänken.",
-                }
-            ),
-            500,
-        )
-    functions.log_admin_action(
-        admin_name,
-        "skapade företagskonto",
-        f"email_ref={mask_email_reference(normalized_email)}",
-    )
-    logger.info(
-        "Admin created supervisor %s",
-        mask_email_reference(normalized_email),
-        extra={"admin": admin_name},
-    )
-    return jsonify({"status": "success", "message": "Företagskonto skapat.", "link": link})
-
-
-@app.post("/admin/api/foretagskonto/koppla")
-def admin_link_supervisor_route():  # pragma: no cover
-    admin_name = _require_admin()
-    payload = request.get_json(silent=True) or {}
-    orgnr = (payload.get("orgnr") or "").strip()
-    personnummer = (payload.get("personnummer") or "").strip()
-    if not orgnr or not personnummer:
-        logger.debug(
-            "Admin link_supervisor without orgnr or personnummer", extra={"admin": admin_name}
-        )
-        return jsonify(
-            {"status": "error", "message": "Ange organisationsnummer och personnummer."}
-        ), 400
-
-    try:
-        success, reason, email_hash = functions.admin_link_supervisor_to_user(orgnr, personnummer)
-    except ValueError:
-        personnummer_masked = mask_hash(functions.hash_value(personnummer))
-        logger.debug(
-            "Admin link_supervisor with invalid orgnr or personnummer hash: %s, %s",
-            orgnr,
-            personnummer_masked,
-            extra={"admin": admin_name},
-        )
-        return jsonify({"status": "error", "message": "Ogiltiga uppgifter."}), 400
-
-    if not success:
-        status_code = 400
-        message = "Åtgärden kunde inte utföras."
-        if reason == "missing_supervisor":
-            status_code = 404
-            message = "Företagskontot finns inte."
-        elif reason == "missing_user":
-            status_code = 404
-            message = "Standardkontot finns inte."
-        elif reason == "exists":
-            status_code = 409
-            message = "Kopplingen finns redan."
-        logger.debug("Admin link_supervisor failed: %s", reason, extra={"admin": admin_name})
-        return jsonify({"status": "error", "message": message}), status_code
-
-    if not email_hash:
-        logger.debug(
-            "Admin link_supervisor missing email hash for orgnr %s",
-            orgnr,
-            extra={"admin": admin_name},
-        )
-        return jsonify({"status": "error", "message": "Företagskontot hittades inte."}), 404
-    normalized_orgnr = functions.validate_orgnr(orgnr)
-    normalized_personnummer = functions.normalize_personnummer(personnummer)
-    personnummer_hash = functions.hash_value(normalized_personnummer)
-    functions.log_admin_action(
-        admin_name,
-        "kopplade företagskonto",
-        (
-            f"orgnr={normalized_orgnr}, email_ref={mask_email_reference(email_hash)}, "
-            f"personnummer_hash={personnummer_hash}"
-        ),
-    )
-    logger.info(
-        "Admin linked supervisor %s to user %s",
-        mask_email_reference(email_hash),
-        mask_hash(personnummer_hash),
-        extra={"admin": admin_name},
-    )
-    return jsonify(
-        {"status": "success", "message": "Företagskontot har kopplats till standardkontot."}
-    )
-
-
-@app.post("/admin/api/foretagskonto/oversikt")
-def admin_supervisor_overview():  # pragma: no cover
-    admin_name = _require_admin()
-    payload = request.get_json(silent=True) or {}
-    orgnr = (payload.get("orgnr") or "").strip()
-    if not orgnr:
-        logger.debug("Admin supervisor_overview without orgnr", extra={"admin": admin_name})
-        return jsonify({"status": "error", "message": "Ange organisationsnummer."}), 400
-
-    try:
-        details = functions.get_supervisor_login_details_for_orgnr(orgnr)
-    except ValueError:
-        logger.debug(
-            "Admin supervisor_overview with invalid orgnr: %s", orgnr, extra={"admin": admin_name}
-        )
-        return jsonify({"status": "error", "message": "Ogiltigt organisationsnummer."}), 400
-
-    if not details:
-        logger.debug(
-            "Admin supervisor_overview not found for orgnr: %s", orgnr, extra={"admin": admin_name}
-        )
-        return jsonify({"status": "error", "message": "Företagskontot hittades inte."}), 404
-
-    email_hash = details["email_hash"]
-    overview = functions.get_supervisor_overview(email_hash)
-    if not overview:
-        logger.debug(
-            "Admin supervisor_overview not found for email reference: %s",
-            mask_email_reference(email_hash),
-            extra={"admin": admin_name},
-        )
-        return jsonify({"status": "error", "message": "Företagskontot hittades inte."}), 404
-
-    functions.log_admin_action(
-        admin_name,
-        "visade företagskontoöversikt",
-        f"orgnr={details['orgnr']}, email_ref={mask_email_reference(email_hash)}",
-    )
-    logger.debug(
-        "Admin supervisor_overview for %s with %d users",
-        mask_email_reference(email_hash),
-        len(overview.get("users", [])),
-        extra={"admin": admin_name},
-    )
-    return jsonify({"status": "success", "data": overview})
-
-
-@app.post("/admin/api/foretagskonto/ta-bort")
-def admin_remove_supervisor_connection():  # pragma: no cover
-    admin_name = _require_admin()
-    if not validate_csrf_token():
-        return jsonify({"status": "error", "message": "Ogiltig CSRF-token."}), 403
-    payload = request.get_json(silent=True) or {}
-    orgnr = (payload.get("orgnr") or "").strip()
-    personnummer = (payload.get("personnummer") or "").strip()
-    personnummer_hash = (payload.get("personnummer_hash") or "").strip()
-    if not orgnr or (not personnummer and not personnummer_hash):
-        return jsonify(
-            {
-                "status": "error",
-                "message": "Ange organisationsnummer och personnummer eller hash.",
-            }
-        ), 400
-
-    try:
-        details = functions.get_supervisor_login_details_for_orgnr(orgnr)
-        if personnummer:
-            normalized_personnummer = functions.normalize_personnummer(personnummer)
-            personnummer_hash = functions.hash_value(normalized_personnummer)
-        elif not functions._is_valid_hash(personnummer_hash):
-            raise ValueError("Ogiltig hash")
-    except ValueError:
-        return jsonify({"status": "error", "message": "Ogiltiga uppgifter."}), 400
-
-    if not details:
-        return jsonify({"status": "error", "message": "Företagskontot hittades inte."}), 404
-
-    if not functions.supervisor_has_access(details["email_hash"], personnummer_hash):
-        return jsonify({"status": "error", "message": "Kopplingen hittades inte."}), 404
-
-    if not functions.supervisor_remove_connection(details["email_hash"], personnummer_hash):
-        return jsonify({"status": "error", "message": "Kunde inte ta bort kopplingen."}), 400
-
-    functions.log_admin_action(
-        admin_name,
-        "tog bort företagskoppling",
-        (
-            f"orgnr={details['orgnr']}, "
-            f"email_ref={mask_email_reference(details['email_hash'])}, "
-            f"personnummer_hash={personnummer_hash}"
-        ),
-    )
-    logger.info(
-        "Admin removed supervisor connection for %s",
-        mask_hash(personnummer_hash),
-        extra={"admin": admin_name},
-    )
-    return jsonify({"status": "success", "message": "Kopplingen har tagits bort."})
-
-
-@app.post("/admin/api/foretagskonto/uppdatera-koppling")
-def admin_change_supervisor_connection():  # pragma: no cover
-    admin_name = _require_admin()
-    if not validate_csrf_token():
-        return jsonify({"status": "error", "message": "Ogiltig CSRF-token."}), 403
-    payload = request.get_json(silent=True) or {}
-    from_orgnr = (payload.get("from_orgnr") or "").strip()
-    to_orgnr = (payload.get("to_orgnr") or "").strip()
-    personnummer = (payload.get("personnummer") or "").strip()
-    if not from_orgnr or not to_orgnr or not personnummer:
-        return jsonify(
-            {
-                "status": "error",
-                "message": "Ange nuvarande organisationsnummer, nytt organisationsnummer och personnummer.",
-            }
-        ), 400
-
-    try:
-        from_details = functions.get_supervisor_login_details_for_orgnr(from_orgnr)
-        to_details = functions.get_supervisor_login_details_for_orgnr(to_orgnr)
-        normalized_personnummer = functions.normalize_personnummer(personnummer)
-    except ValueError:
-        return jsonify({"status": "error", "message": "Ogiltiga uppgifter."}), 400
-
-    if not from_details or not to_details:
-        return jsonify({"status": "error", "message": "Företagskontot hittades inte."}), 404
-
-    personnummer_hash = functions.hash_value(normalized_personnummer)
-    if not functions.supervisor_has_access(from_details["email_hash"], personnummer_hash):
-        return jsonify({"status": "error", "message": "Den gamla kopplingen hittades inte."}), 404
-
-    if not functions.supervisor_remove_connection(from_details["email_hash"], personnummer_hash):
-        return jsonify(
-            {"status": "error", "message": "Kunde inte ta bort den gamla kopplingen."}
-        ), 400
-
-    success, reason, email_hash = functions.admin_link_supervisor_to_user(
-        to_orgnr, normalized_personnummer
-    )
-    if not success:
-        rollback_success, rollback_reason, rollback_email_hash = (
-            functions.admin_link_supervisor_to_user(from_orgnr, normalized_personnummer)
-        )
-        message = "Kunde inte skapa den nya kopplingen."
-        if reason == "missing_supervisor":
-            message = "Det nya företagskontot hittades inte."
-        elif reason == "missing_user":
-            message = "Standardkontot hittades inte."
-        elif reason == "exists":
-            message = "Det finns redan en koppling till det nya företagskontot."
-        if not rollback_success:
-            logger.error(
-                (
-                    "Rollback misslyckades vid ändring av koppling: "
-                    "reason=%s, rollback_reason=%s, rollback_email_ref=%s"
-                ),
-                reason,
-                rollback_reason,
-                mask_email_reference(rollback_email_hash or ""),
-            )
-            return jsonify(
-                {
-                    "status": "error",
-                    "message": (f"{message} Rollback misslyckades: {rollback_reason}."),
-                }
-            ), 500
-        return jsonify({"status": "error", "message": message}), 400
-
-    functions.log_admin_action(
-        admin_name,
-        "ändrade företagskoppling",
-        (
-            f"from_orgnr={from_details['orgnr']}, to_orgnr={to_details['orgnr']}, "
-            f"email_ref={mask_email_reference(email_hash or '')}, "
-            f"personnummer_hash={personnummer_hash}"
-        ),
-    )
-    logger.info(
-        "Admin changed supervisor connection for %s",
-        mask_hash(personnummer_hash),
-        extra={"admin": admin_name},
-    )
-    return jsonify({"status": "success", "message": "Kopplingen har uppdaterats."})
-
-
-@app.post("/admin/api/foretagskonto/radera")
-def admin_delete_supervisor_account_route():  # pragma: no cover
-    admin_name = _require_admin()
-    if not validate_csrf_token():
-        return jsonify({"status": "error", "message": "Ogiltig CSRF-token."}), 403
-    payload = request.get_json(silent=True) or {}
-    orgnr = (payload.get("orgnr") or "").strip()
-    if not orgnr:
-        return jsonify({"status": "error", "message": "Ange organisationsnummer."}), 400
-
-    try:
-        deleted, summary, normalized_orgnr = functions.admin_delete_supervisor_account(orgnr)
-    except ValueError:
-        return jsonify({"status": "error", "message": "Ogiltigt organisationsnummer."}), 400
-    except Exception:
-        logger.error("Misslyckades att radera företagskonto för %s", orgnr)
-        return jsonify({"status": "error", "message": "Kunde inte radera företagskontot."}), 500
-
-    if not deleted:
-        return jsonify({"status": "error", "message": "Företagskontot hittades inte."}), 404
-
-    summary_details = ", ".join(f"{key}={value}" for key, value in summary.items())
-    functions.log_admin_action(
-        admin_name,
-        "raderade företagskonto",
-        f"orgnr={normalized_orgnr}, {summary_details}",
-    )
-    logger.info(
-        "Admin deleted supervisor account for %s",
-        mask_hash(functions.hash_value(normalized_orgnr)),
-        extra={"admin": admin_name},
-    )
-    return jsonify(
-        {
-            "status": "success",
-            "message": "Företagskontot har raderats.",
-            "data": summary,
-        }
-    )
-
-
-@app.get("/admin/avancerat")
-def admin_advanced():  # pragma: no cover
-    if not session.get("admin_logged_in"):
-        logger.warning("Obehörigt admin advanced-GET-anrop")
-        return redirect("/login_admin")
-    tables = sorted(functions.TABLE_REGISTRY.keys())
-    return render_template("admin_advanced.html", tables=tables)
-
-
-@app.get("/admin/advanced/api/schema/<table_name>")
-def admin_advanced_schema(table_name: str):  # pragma: no cover
-    admin_name = _require_admin()
-    try:
-        schema = functions.get_table_schema(table_name)
-    except ValueError:
-        logger.debug(
-            "Admin advanced schema with unknown table: %s",
-            table_name,
-            extra={"admin": admin_name},
-        )
-        return jsonify({"status": "error", "message": "Okänd tabell."}), 404
-    logger.debug("Admin advanced schema for table: %s", table_name, extra={"admin": admin_name})
-    return jsonify({"status": "success", "schema": schema})
-
-
-@app.get("/admin/advanced/api/rows/<table_name>")
-def admin_advanced_rows(table_name: str):  # pragma: no cover
-    admin_name = _require_admin()
-    search_term = request.args.get("sok")
-    limit = request.args.get("limit", type=int) or 100
-    try:
-        rows = functions.fetch_table_rows(table_name, search_term, limit)
-    except ValueError:
-        logger.debug(
-            "Admin advanced rows with unknown table: %s",
-            table_name,
-            extra={"admin": admin_name},
-        )
-        return jsonify({"status": "error", "message": "Okänd tabell."}), 404
-    masked_search = _sanitize_search_term(search_term)
-    logger.debug(
-        "Admin advanced rows for table: %s, search: %r, limit: %d",
-        table_name,
-        masked_search,
-        limit,
-        extra={"admin": admin_name},
-    )
-    return jsonify({"status": "success", "rows": rows})
-
-
-@app.post("/admin/advanced/api/rows/<table_name>")
-def admin_advanced_create(table_name: str):  # pragma: no cover
-    admin_name = _require_admin()
-    values = request.get_json(silent=True) or {}
-    try:
-        row = functions.create_table_row(table_name, values)
-    except ValueError as exc:
-        logger.warning("Fel i create_table_row: %s", exc)
-        return jsonify({"status": "error", "message": "Kunde inte skapa posten."}), 400
-    masked_row = _mask_sensitive_fields(row)
-    functions.log_admin_action(
-        admin_name,
-        "skapade post",
-        f"tabell={table_name}",
-    )
-    logger.info(
-        "Admin created row in table %s",
-        table_name,
-        extra={"admin": admin_name},
-    )
-    logger.debug(
-        "Admin created masked row in table %s: %s",
-        table_name,
-        masked_row,
-        extra={"admin": admin_name},
-    )
-    return jsonify({"status": "success", "row": masked_row}), 201
-
-
-@app.put("/admin/advanced/api/rows/<table_name>/<int:row_id>")
-def admin_advanced_update(table_name: str, row_id: int):  # pragma: no cover
-    admin_name = _require_admin()
-    values = request.get_json(silent=True) or {}
-    try:
-        updated = functions.update_table_row(table_name, row_id, values)
-    except ValueError as exc:
-        logger.error(
-            "Misslyckades med att uppdatera rad i tabell '%s', id=%s: %s",
-            table_name,
-            row_id,
-            exc,
-            exc_info=True,
-        )
-        return jsonify({"status": "error", "message": "Felaktiga data."}), 400
-    if not updated:
-        logger.debug("Admin advanced update with missing row: table=%s, id=%d", table_name, row_id)
-        return jsonify({"status": "error", "message": "Posten hittades inte."}), 404
-    functions.log_admin_action(
-        admin_name,
-        "uppdaterade post",
-        f"tabell={table_name}, id={row_id}",
-    )
-    logger.info(
-        "Admin updated row in table %s, id=%d: %s",
-        table_name,
-        row_id,
-        _mask_sensitive_fields(values),
-        extra={"admin": admin_name},
-    )
-    return jsonify({"status": "success"})
-
-
-@app.delete("/admin/advanced/api/rows/<table_name>/<int:row_id>")
-def admin_advanced_delete(table_name: str, row_id: int):  # pragma: no cover
-    admin_name = _require_admin()
-    try:
-        deleted = functions.delete_table_row(table_name, row_id)
-    except ValueError:
-        logger.debug("Admin advanced delete with unknown table: %s", table_name)
-        return jsonify({"status": "error", "message": "Okänd tabell."}), 404
-    if not deleted:
-        logger.debug("Admin advanced delete with missing row: table=%s, id=%d", table_name, row_id)
-        return jsonify({"status": "error", "message": "Posten hittades inte."}), 404
-    functions.log_admin_action(
-        admin_name,
-        "raderade post",
-        f"tabell={table_name}, id={row_id}",
-    )
-    logger.info(
-        "Admin deleted row in table %s, id=%d", table_name, row_id, extra={"admin": admin_name}
-    )
-    return jsonify({"status": "success"})
-
-
-@app.route("/verify_certificate/<personnummer>", methods=["GET"])
-def verify_certificate_route(personnummer):  # pragma: no cover
-    # Allow an admin to verify whether a user's certificate is confirmed.
-    #
-    # Uses a cached lookup to avoid repeated database queries for the same
-    # ``personnummer``. Returns a JSON response indicating the verification
-    # status. If the certificate isn't verified, an informative message is sent
-    # back to the administrator.
-    if not session.get("admin_logged_in"):
-        logger.warning("Obehörigt försök att verifiera intyg")
-        return redirect("/login_admin")
-
-    if functions.verify_certificate(personnummer):
-        logger.info("Intyg för %s är verifierat", mask_hash(functions.hash_value(personnummer)))
-        return jsonify({"status": "success", "verified": True})
-    logger.info("Intyg för %s är INTE verifierat", mask_hash(functions.hash_value(personnummer)))
-    return jsonify(
-        {
-            "status": "error",
-            "message": "Standardkontots certifikat är inte verifierat",
-        }
-    ), 404
-
-
-@app.route("/login_admin", methods=["POST", "GET"])
-def login_admin():  # pragma: no cover
-    # Authenticate an administrator for access to the admin panel.
-    if request.method == "POST":
-        admin_password = os.getenv("admin_password")
-        admin_username = os.getenv("admin_username")
-
-        # Require admin credentials to be explicitly set (no insecure defaults)
-        if not admin_password or not admin_username:
-            error_msg = "KRITISKT: miljövariablerna admin_username och admin_password måste vara satta och inte tomma"
-            logger.critical(error_msg)
-            critical_events.send_critical_error_notification(error_message=error_msg, endpoint="/login_admin", user_ip=get_request_ip())
-            raise RuntimeError(error_msg)
-        submitted_username = request.form.get("username")
-        submitted_password = request.form.get("password")
-        user_ok = secrets.compare_digest(str(submitted_username or ""), str(admin_username))
-        pass_ok = secrets.compare_digest(str(submitted_password or ""), str(admin_password))
-        if user_ok and pass_ok:
-            session["admin_logged_in"] = True
-            session["admin_username"] = admin_username
-            logger.info("Admin %s loggade in", _mask_username_for_log(admin_username))
-            return redirect("/admin")
-
-        logger.warning(
-            "Ogiltigt admin-inloggningsförsök för %s",
-            _mask_username_for_log(submitted_username),
-        )
-        return jsonify({"status": "error", "message": "Ogiltiga inloggningsuppgifter"})
-
-    elif request.method == "GET":
-        logger.debug("Renderar admin-inloggningssida")
-        return render_template("admin_login.html")
-    else:
-        logger.warning("Ogiltig förfrågningsmetod %s till login_admin", request.method)
-        return jsonify(
-            {"status": "error", "message": "Ogiltig HTTP-metod", "method": request.method}
-        )
-
-
-@app.route("/logout")
-def logout():
-    # Logga ut både admin och användare.
-    logger.info("Loggar ut användare och admin")
-    session.pop("user_logged_in", None)
-    session.pop("admin_logged_in", None)
-    session.pop("admin_username", None)
-    session.pop("personnummer", None)
-    session.pop("personnummer_raw", None)
-    session.pop("supervisor_logged_in", None)
-    session.pop("supervisor_email_hash", None)
-    session.pop("supervisor_name", None)
-    session.pop("supervisor_orgnr", None)
-    return redirect("/")
-
-
-## -------------------------Error Handlers -------------------------##
-@app.route("/error")
-def error():  # pragma: no cover
-    # Intentionally raise an error to test the 500 page.
-    # This will cause a 500 Internal Server Error
-    raise Exception("Testing 500 error page")
-
-
-@app.errorhandler(500)
-def internal_server_error(_):  # pragma: no cover
-    logger.error("500 Internt serverfel: %s", request.path)
-
-    try:
-        user_ip = get_request_ip()
-        endpoint = request.path
-        error_msg = f"Endpoint: {endpoint}\nMetod: {request.method}\nIP: {user_ip}"
-        critical_events.send_critical_error_notification(
-            error_message=error_msg, endpoint=endpoint, user_ip=user_ip
-        )
-    except Exception as e:
-        logger.warning("Kunde inte skicka error-notifikation: %s", e)
-
-    # Visa en användarvänlig 500-sida när ett serverfel inträffar.
-    error_code = 500
-    error_message = "Ett internt serverfel har inträffat. Vänligen försök igen senare."
-    return render_template(
-        "error.html", error_code=error_code, error_message=error_message, time=time.time()
-    ), 500
-
-
-@app.errorhandler(Exception)
-def handle_unexpected_exception(error: Exception):  # pragma: no cover
-    # Logga oväntade fel och låt HTTP-fel hanteras av sina egna handlers.
-    if isinstance(error, HTTPException):
-        return error
-    logger.error("Oväntat fel inträffade: %s", str(error))
-    return internal_server_error(error)
-
-
-@app.errorhandler(401)
-def unauthorized_error(_):  # pragma: no cover
-    # Visa en användarvänlig 401-sida vid obehörig åtkomst.
-    logger.warning("401 Obehörig åtkomst: %s", request.path)
-    error_code = 401
-    error_message = "Du måste vara inloggad för att se denna sida."
-    return render_template(
-        "error.html", error_code=error_code, error_message=error_message, time=time.time()
-    ), 401
-
-
-@app.errorhandler(409)
-def conflict_error(_):  # pragma: no cover
-    # Visa en användarvänlig 409-sida vid konflikt.
-    logger.error("409 Konflikt: %s", request.path)
-    error_code = 409
-    error_message = "Det uppstod en konflikt vid hantering av din begäran."
-    return render_template(
-        "error.html", error_code=error_code, error_message=error_message, time=time.time()
-    ), 409
-
-
-@app.errorhandler(413)
-def request_entity_too_large(_):  # pragma: no cover
-    # Visa ett tydligt felmeddelande när uppladdningen överskrider global gräns.
-    logger.warning("413 För stor uppladdning: %s", request.path)
-    endpoint = request.endpoint or ""
-    if endpoint.startswith("admin") or request.path == "/admin" or request.path.startswith(
-        "/admin/"
-    ):
-        return jsonify({"status": "error", "message": UPLOAD_TOO_LARGE_MESSAGE}), 413
-    if endpoint in {"user_upload_page", "user_upload_pdf_route"} or request.path in {
-        "/dashboard/ladda-upp",
-        "/dashboard/upload",
-    } or request.path.startswith("/dashboard/ladda-upp/") or request.path.startswith(
-        "/dashboard/upload/"
-    ):
-        flash(UPLOAD_TOO_LARGE_MESSAGE, "error")
-        return redirect(url_for("user_upload_page"))
-
-    error_code = 413
-    error_message = UPLOAD_TOO_LARGE_MESSAGE
-    return render_template(
-        "error.html", error_code=error_code, error_message=error_message, time=time.time()
-    ), 413
-
-
-@app.errorhandler(404)
-def page_not_found(_):  # pragma: no cover
-    # Visa en användarvänlig 404-sida när en sida saknas.
-    logger.warning("Sidan hittades inte: %s", request.path)
-    error_code = 404
-    error_message = "Sidan du letade efter kunde inte hittas."
-    return render_template(
-        "error.html", error_code=error_code, error_message=error_message, time=time.time()
-    ), 404
-
-
-##----------------------------------------##
-
-
-@app.template_filter("datetimeformat")
-def datetimeformat(value, format="%Y-%m-%d %H:%M:%S"):  # pragma: no cover
-    # Format a POSIX timestamp for display in templates.
-    import datetime
-
-    return datetime.datetime.fromtimestamp(value).strftime(format)
 
 
 if __name__ == "__main__":  # pragma: no cover
@@ -4264,17 +153,22 @@ if __name__ == "__main__":  # pragma: no cover
     except KeyboardInterrupt:
         logger.info("Applikationen avbröts av användaren")
         try:
-            critical_events.send_crash_notification("Application interrupted by user (KeyboardInterrupt)")
-        except Exception as e:
-            logger.critical("Misslyckades med att skicka kraschnotifikation: %s", e)
-            critical_events.send_crash_notification("Misslyckades med att skicka kraschnotifikation: " + str(e))
-    except Exception as e:
-        logger.critical("Applikationen kraschade med undantag: %s", e, exc_info=True)
+            critical_events.send_crash_notification(
+                "Application interrupted by user (KeyboardInterrupt)"
+            )
+        except Exception as exc:
+            logger.critical("Misslyckades med att skicka kraschnotifikation: %s", exc)
+            critical_events.send_crash_notification(
+                "Misslyckades med att skicka kraschnotifikation: " + str(exc)
+            )
+    except Exception as exc:
+        logger.critical("Applikationen kraschade med undantag: %s", exc, exc_info=True)
         try:
-            error_details = f"Exception: {type(e).__name__}\nMessage: {str(e)}"
+            error_details = f"Exception: {type(exc).__name__}\nMessage: {str(exc)}"
             critical_events.send_crash_notification(error_details)
         except Exception as alert_error:
             logger.critical("Misslyckades med att skicka kraschnotifikation: %s", alert_error)
         raise
 
-# © 2026 Liam Suorsa. All rights reserved.
+
+# Copyright (c) Liam Suorsa and Mika Suorsa
