+++ web/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) Liam Suorsa and Mika Suorsa
+"""Internt webblager för Flask-appens routes, bootstrap och felhantering."""
+
+# Copyright (c) Liam Suorsa and Mika Suorsa
